Split the command handlers and class definitions out of ops
pmix_server_ops.c was 3700 lines holding two unrelated things: every
per-command handler the switchyard dispatches, and every
PMIX_CLASS_INSTANCE the server role defines. Neither was easy to
find inside the other.

Take the class definitions out first - they are a self-contained
block and belong together, because reading a destructor is how you
learn what a type owns. pmix_server_classes.c now holds all of them.

Then group the handlers by command family:

  pmix_server_ops.c        abort, publish/lookup/unpublish, spawn,
                           and the pmix_host_server definition
  pmix_server_events.c     the event family - registration, the
                           forwarding bookkeeping, cached-event
                           replay, notify-from-client
  pmix_server_control.c    query, log, allocate, job control,
                           monitor, credentials, session control,
                           resource blocks, cache refresh
  pmix_server_fabric.c     fabric register/update, device distances

Two groups go home rather than into a new file. The IOF commands -
iofreg, iofdereg, iofstdin and pmix_server_process_iof - join
pmix_server_iof.c, which already owned the forwarding side. And
pmix_server_trk_complete and pmix_server_set_collective_status join
pmix_server_fence.c, which owns the collective-tracker engine those
two predicates belong to; the doc has said so for a while.

Nothing else changed. Every line of the old file is accounted for in
the new set, verbatim, and no symbol had to change visibility - the
statics all turned out to be confined to the family that used them.
