+++ src/server/pmix_server_ops.c
@@ -496,44 +496,6 @@ cleanup:
     return rc;
 }
 
-pmix_status_t pmix_server_process_iof(pmix_setup_caddy_t *cd,
-                                      char nspace[])
-{
-    pmix_iof_req_t *req;
-    pmix_status_t rc;
-    pmix_iof_cache_t *iof, *ionext;
-
-    // if no channels to forward, just return success
-    if (PMIX_FWD_NO_CHANNELS == cd->channels) {
-        return PMIX_SUCCESS;
-    }
-
-    /* record the request */
-    req = PMIX_NEW(pmix_iof_req_t);
-    if (NULL == req) {
-        return PMIX_ERR_NOMEM;
-    }
-    PMIX_RETAIN(cd->peer);
-    req->requestor = cd->peer;
-    req->nprocs = 1;
-    PMIX_PROC_CREATE(req->procs, req->nprocs);
-    PMIX_LOAD_PROCID(&req->procs[0], nspace, PMIX_RANK_WILDCARD);
-    req->channels = cd->channels;
-    req->flags = cd->flags;
-    req->local_id = pmix_pointer_array_add(&pmix_globals.iof_requests, req);
-    /* process any cached IO */
-    PMIX_LIST_FOREACH_SAFE (iof, ionext, &pmix_server_globals.iof, pmix_iof_cache_t) {
-        rc = pmix_iof_process_iof(iof->channel, &iof->source, iof->bo,
-                                  iof->info, iof->ninfo, req);
-        if (PMIX_OPERATION_SUCCEEDED == rc) {
-            /* remove it from the list since it has now been forwarded */
-            pmix_list_remove_item(&pmix_server_globals.iof, &iof->super);
-            PMIX_RELEASE(iof);
-        }
-    }
-    return PMIX_SUCCESS;
-}
-
 static void _spcbfunc(int sd, short args, void *cbdata)
 {
     pmix_setup_caddy_t *cd = (pmix_setup_caddy_t *) cbdata;
@@ -769,2944 +731,3 @@ cleanup:
     }
     return rc;
 }
-
-static void _check_cached_events(int sd, short args, void *cbdata)
-{
-    pmix_setup_caddy_t *scd = (pmix_setup_caddy_t *) cbdata;
-    pmix_notify_caddy_t *cd;
-    pmix_range_trkr_t rngtrk;
-    pmix_proc_t proc;
-    int i;
-    size_t k, n;
-    bool found, matched;
-    pmix_buffer_t *relay;
-    pmix_status_t ret = PMIX_SUCCESS;
-    pmix_cmd_t cmd = PMIX_NOTIFY_CMD;
-
-    PMIX_HIDE_UNUSED_PARAMS(sd, args);
-
-    /* check if any matching notifications have been cached */
-    rngtrk.procs = NULL;
-    rngtrk.nprocs = 0;
-    for (i = 0; i < pmix_globals.max_events; i++) {
-        pmix_hotel_knock(&pmix_globals.notifications, i, (void **) &cd);
-        if (NULL == cd) {
-            continue;
-        }
-        found = false;
-        if (NULL == scd->codes) {
-            if (!cd->nondefault) {
-                /* they registered a default event handler - always matches */
-                found = true;
-            }
-        } else {
-            for (k = 0; k < scd->ncodes; k++) {
-                if (scd->codes[k] == cd->status) {
-                    found = true;
-                    break;
-                }
-            }
-        }
-        if (!found) {
-            continue;
-        }
-        /* check if the affected procs (if given) match those they
-         * wanted to know about */
-        if (!pmix_notify_check_affected(cd->affected, cd->naffected, scd->procs, scd->nprocs)) {
-            continue;
-        }
-        /* check the range */
-        if (NULL == cd->targets) {
-            rngtrk.procs = &cd->source;
-            rngtrk.nprocs = 1;
-        } else {
-            rngtrk.procs = cd->targets;
-            rngtrk.nprocs = cd->ntargets;
-        }
-        rngtrk.range = cd->range;
-        PMIX_LOAD_PROCID(&proc, scd->peer->info->pname.nspace, scd->peer->info->pname.rank);
-        if (!pmix_notify_check_range(&rngtrk, &proc)) {
-            continue;
-        }
-        /* if we were given specific targets, check if this is one */
-        found = false;
-        if (NULL != cd->targets) {
-            matched = false;
-            for (n = 0; n < cd->ntargets; n++) {
-                /* if the source of the event is the same peer just registered, then ignore it
-                 * as the event notification system will have already locally
-                 * processed it */
-                if (PMIX_CHECK_NAMES(&cd->source, &scd->peer->info->pname)) {
-                    continue;
-                }
-                if (PMIX_CHECK_NAMES(&scd->peer->info->pname, &cd->targets[n])) {
-                    matched = true;
-                    /* track the number of targets we have left to notify */
-                    --cd->nleft;
-                    /* if this is the last one, then evict this event
-                     * from the cache */
-                    if (0 == cd->nleft) {
-                        pmix_hotel_checkout(&pmix_globals.notifications, cd->room);
-                        found = true; // mark that we should release cd
-                    }
-                    break;
-                }
-            }
-            if (!matched) {
-                /* do not notify this one */
-                continue;
-            }
-        }
-
-        /* all matches - notify */
-        relay = PMIX_NEW(pmix_buffer_t);
-        if (NULL == relay) {
-            /* nothing we can do */
-            PMIX_ERROR_LOG(PMIX_ERR_NOMEM);
-            ret = PMIX_ERR_NOMEM;
-            if (found) {
-                /* this event was checked out of the hotel above */
-                PMIX_RELEASE(cd);
-            }
-            break;
-        }
-        /* pack the info data stored in the event */
-        PMIX_BFROPS_PACK(ret, scd->peer, relay, &cmd, 1, PMIX_COMMAND);
-        if (PMIX_SUCCESS == ret) {
-            PMIX_BFROPS_PACK(ret, scd->peer, relay, &cd->status, 1, PMIX_STATUS);
-        }
-        if (PMIX_SUCCESS == ret) {
-            PMIX_BFROPS_PACK(ret, scd->peer, relay, &cd->source, 1, PMIX_PROC);
-        }
-        if (PMIX_SUCCESS == ret) {
-            PMIX_BFROPS_PACK(ret, scd->peer, relay, &cd->ninfo, 1, PMIX_SIZE);
-        }
-        if (PMIX_SUCCESS == ret && 0 < cd->ninfo) {
-            PMIX_BFROPS_PACK(ret, scd->peer, relay, cd->info, cd->ninfo, PMIX_INFO);
-        }
-        if (PMIX_SUCCESS != ret) {
-            /* release the relay we could not fill, and the event if it was
-             * checked out of the hotel, before abandoning the loop */
-            PMIX_ERROR_LOG(ret);
-            PMIX_RELEASE(relay);
-            if (found) {
-                PMIX_RELEASE(cd);
-            }
-            break;
-        }
-        PMIX_SERVER_QUEUE_REPLY(ret, scd->peer, 0, relay);
-        if (PMIX_SUCCESS != ret) {
-            PMIX_RELEASE(relay);
-        }
-        if (found) {
-            PMIX_RELEASE(cd);
-        }
-    }
-    /* release the caddy */
-    if (NULL != scd->codes) {
-        free(scd->codes);
-    }
-    if (NULL != scd->info) {
-        PMIX_INFO_FREE(scd->info, scd->ninfo);
-    }
-    if (NULL != scd->opcbfunc) {
-        scd->opcbfunc(ret, scd->cbdata);
-    }
-    PMIX_RELEASE(scd);
-}
-
-/* provide a callback function for the host when it finishes
- * processing the registration */
-static void regevopcbfunc(pmix_status_t status, void *cbdata)
-{
-    pmix_setup_caddy_t *cd = (pmix_setup_caddy_t *) cbdata;
-
-    if (pmix_atomic_check_bool(&pmix_globals.progress_thread_stopped)) {
-        PMIX_RELEASE(cd);
-        return;
-    }
-
-    /* if the registration succeeded, then check local cache */
-    if (PMIX_SUCCESS == status) {
-        cd->status = status;
-        PMIX_THREADSHIFT(cd, _check_cached_events);
-        return;
-    }
-
-    /* it didn't succeed, so cleanup and execute the callback
-     * so we don't hang */
-    if (NULL != cd->codes) {
-        free(cd->codes);
-    }
-    if (NULL != cd->info) {
-        PMIX_INFO_FREE(cd->info, cd->ninfo);
-    }
-    if (NULL != cd->opcbfunc) {
-        cd->opcbfunc(status, cd->cbdata);
-    }
-    PMIX_RELEASE(cd);
-}
-
-/* the host is done processing our request to stop forwarding a code -
- * all we carried across the up-call was the array of codes itself */
-static void deregevopcbfunc(pmix_status_t status, void *cbdata)
-{
-    pmix_status_t *codes = (pmix_status_t *) cbdata;
-
-    PMIX_HIDE_UNUSED_PARAMS(status);
-
-    free(codes);
-}
-
-/* ask our host to stop forwarding the given event codes - takes
- * ownership of the (malloc'd) codes array */
-static void stop_forwarding(pmix_status_t *codes, size_t ncodes)
-{
-    pmix_status_t rc;
-
-    if (NULL == pmix_host_server.deregister_events) {
-        free(codes);
-        return;
-    }
-    pmix_output_verbose(2, pmix_server_globals.event_output,
-                        "server deregister events: telling host to stop forwarding %lu code(s)",
-                        (unsigned long) ncodes);
-    rc = pmix_host_server.deregister_events(codes, ncodes, deregevopcbfunc, codes);
-    if (PMIX_SUCCESS != rc) {
-        /* the host either completed the operation itself or rejected
-         * it - either way, no callback is coming */
-        free(codes);
-    }
-}
-
-bool pmix_server_prune_reginfo(pmix_regevents_info_t *reginfo)
-{
-    pmix_status_t *codes;
-
-    if (0 < pmix_list_get_size(&reginfo->peers) || 0 < reginfo->nmine) {
-        /* somebody still wants this code */
-        return false;
-    }
-
-    pmix_list_remove_item(&pmix_server_globals.events, &reginfo->super);
-    if (reginfo->active) {
-        codes = (pmix_status_t *) malloc(sizeof(pmix_status_t));
-        if (NULL != codes) {
-            codes[0] = reginfo->code;
-            stop_forwarding(codes, 1);
-        }
-    }
-    PMIX_RELEASE(reginfo);
-    return true;
-}
-
-/* find the tracking object for the given code, creating it if necessary */
-static pmix_regevents_info_t *get_reginfo(pmix_status_t code)
-{
-    pmix_regevents_info_t *reginfo;
-
-    PMIX_LIST_FOREACH (reginfo, &pmix_server_globals.events, pmix_regevents_info_t) {
-        if (code == reginfo->code) {
-            return reginfo;
-        }
-    }
-    reginfo = PMIX_NEW(pmix_regevents_info_t);
-    if (NULL == reginfo) {
-        return NULL;
-    }
-    reginfo->code = code;
-    pmix_list_append(&pmix_server_globals.events, &reginfo->super);
-    return reginfo;
-}
-
-/* Mark the codes our host is not already forwarding, sorting them to
- * the front of the array, and return their number. The array is treated
- * as an unordered set by everything downstream, so the reordering only
- * lets us hand the host the subset it needs to act upon. */
-static size_t mark_activations(pmix_status_t *codes, size_t ncodes)
-{
-    pmix_regevents_info_t *reginfo;
-    pmix_status_t tmp;
-    size_t n, nactive = 0;
-
-    for (n = 0; n < ncodes; n++) {
-        if (!PMIX_SYSTEM_EVENT(codes[n])) {
-            continue;
-        }
-        PMIX_LIST_FOREACH (reginfo, &pmix_server_globals.events, pmix_regevents_info_t) {
-            if (codes[n] != reginfo->code) {
-                continue;
-            }
-            if (!reginfo->active) {
-                reginfo->active = true;
-                tmp = codes[nactive];
-                codes[nactive] = codes[n];
-                codes[n] = tmp;
-                ++nactive;
-            }
-            break;
-        }
-    }
-    return nactive;
-}
-
-/* our host rejected the request, so it is not forwarding these codes
- * after all - let the next registration ask it again */
-static void undo_activations(pmix_status_t *codes, size_t nactive)
-{
-    pmix_regevents_info_t *reginfo;
-    size_t n;
-
-    for (n = 0; n < nactive; n++) {
-        PMIX_LIST_FOREACH (reginfo, &pmix_server_globals.events, pmix_regevents_info_t) {
-            if (codes[n] == reginfo->code) {
-                reginfo->active = false;
-                break;
-            }
-        }
-    }
-}
-
-size_t pmix_server_activate_events(pmix_status_t *codes, size_t ncodes)
-{
-    pmix_regevents_info_t *reginfo;
-    size_t n;
-
-    if (NULL == codes) {
-        return 0;
-    }
-
-    /* record our own interest in each of the system codes */
-    for (n = 0; n < ncodes; n++) {
-        if (!PMIX_SYSTEM_EVENT(codes[n])) {
-            continue;
-        }
-        reginfo = get_reginfo(codes[n]);
-        if (NULL == reginfo) {
-            PMIX_ERROR_LOG(PMIX_ERR_NOMEM);
-            continue;
-        }
-        ++reginfo->nmine;
-    }
-
-    return mark_activations(codes, ncodes);
-}
-
-void pmix_server_deactivate_events(pmix_status_t *codes, size_t ncodes)
-{
-    pmix_regevents_info_t *reginfo, *regnext;
-    size_t n;
-
-    if (NULL == codes) {
-        return;
-    }
-
-    for (n = 0; n < ncodes; n++) {
-        if (!PMIX_SYSTEM_EVENT(codes[n])) {
-            continue;
-        }
-        PMIX_LIST_FOREACH_SAFE (reginfo, regnext, &pmix_server_globals.events,
-                                pmix_regevents_info_t) {
-            if (codes[n] != reginfo->code) {
-                continue;
-            }
-            if (0 < reginfo->nmine) {
-                --reginfo->nmine;
-            }
-            pmix_server_prune_reginfo(reginfo);
-            break;
-        }
-    }
-}
-
-pmix_status_t pmix_server_register_events(pmix_peer_t *peer, pmix_buffer_t *buf,
-                                          pmix_op_cbfunc_t cbfunc, void *cbdata)
-{
-    int32_t cnt;
-    pmix_status_t rc;
-    pmix_status_t *codes = NULL;
-    pmix_info_t *info = NULL;
-    size_t ninfo = 0, ncodes, n;
-    pmix_regevents_info_t *reginfo, *rptr;
-    pmix_peer_events_info_t *prev = NULL;
-    pmix_setup_caddy_t *scd;
-    bool enviro_events = false;
-    bool found;
-    pmix_proc_t *affected = NULL;
-    size_t naffected = 0;
-    size_t nactive = 0;
-
-    pmix_output_verbose(2, pmix_server_globals.event_output,
-                        "recvd register events for peer %s:%d",
-                        peer->info->pname.nspace, peer->info->pname.rank);
-
-    /* unpack the number of codes */
-    cnt = 1;
-    PMIX_BFROPS_UNPACK(rc, peer, buf, &ncodes, &cnt, PMIX_SIZE);
-    if (PMIX_SUCCESS != rc) {
-        PMIX_ERROR_LOG(rc);
-        return rc;
-    }
-    /* unpack the array of codes */
-    if (0 < ncodes) {
-        codes = (pmix_status_t *) malloc(ncodes * sizeof(pmix_status_t));
-        if (NULL == codes) {
-            rc = PMIX_ERR_NOMEM;
-            goto cleanup;
-        }
-        cnt = ncodes;
-        PMIX_BFROPS_UNPACK(rc, peer, buf, codes, &cnt, PMIX_STATUS);
-        if (PMIX_SUCCESS != rc) {
-            PMIX_ERROR_LOG(rc);
-            goto cleanup;
-        }
-    }
-
-    /* unpack the number of info objects */
-    cnt = 1;
-    PMIX_BFROPS_UNPACK(rc, peer, buf, &ninfo, &cnt, PMIX_SIZE);
-    if (PMIX_SUCCESS != rc) {
-        PMIX_ERROR_LOG(rc);
-        goto cleanup;
-    }
-    /* unpack the array of info objects */
-    if (0 < ninfo) {
-        PMIX_INFO_CREATE(info, ninfo);
-        if (NULL == info) {
-            rc = PMIX_ERR_NOMEM;
-            goto cleanup;
-        }
-        cnt = ninfo;
-        PMIX_BFROPS_UNPACK(rc, peer, buf, info, &cnt, PMIX_INFO);
-        if (PMIX_SUCCESS != rc) {
-            PMIX_ERROR_LOG(rc);
-            goto cleanup;
-        }
-    }
-
-    /* check the directives */
-    for (n = 0; n < ninfo; n++) {
-        if (PMIX_CHECK_KEY(&info[n], PMIX_EVENT_AFFECTED_PROC)) {
-            if (NULL != affected) {
-                PMIX_ERROR_LOG(PMIX_ERR_BAD_PARAM);
-                rc = PMIX_ERR_BAD_PARAM;
-                goto cleanup;
-            }
-            naffected = 1;
-            PMIX_PROC_CREATE(affected, naffected);
-            memcpy(affected, info[n].value.data.proc, sizeof(pmix_proc_t));
-        } else if (PMIX_CHECK_KEY(&info[n], PMIX_EVENT_AFFECTED_PROCS)) {
-            if (NULL != affected) {
-                PMIX_ERROR_LOG(PMIX_ERR_BAD_PARAM);
-                rc = PMIX_ERR_BAD_PARAM;
-                goto cleanup;
-            }
-            naffected = info[n].value.data.darray->size;
-            PMIX_PROC_CREATE(affected, naffected);
-            memcpy(affected, info[n].value.data.darray->array, naffected * sizeof(pmix_proc_t));
-        }
-    }
-
-    /* check the codes for system events */
-    for (n = 0; n < ncodes; n++) {
-        if (PMIX_SYSTEM_EVENT(codes[n])) {
-            enviro_events = true;
-            break;
-        }
-    }
-
-    /* if they asked for enviro events, and our host doesn't support
-     * register_events, then we cannot meet the request */
-    if (enviro_events && NULL == pmix_host_server.register_events) {
-        enviro_events = false;
-        rc = PMIX_ERR_NOT_SUPPORTED;
-        goto cleanup;
-    }
-
-    /* if they didn't send us any codes, then they are registering a
-     * default event handler. In that case, look only at the default
-     * entry - creating it if this is the first default registration we
-     * have seen - and add this request to it.
-     *
-     * Creating it is not optional. This used to walk the list and, if it
-     * found no PMIX_MAX_ERR_CONSTANT entry, fall out of the loop and
-     * return PMIX_OPERATION_SUCCEEDED having recorded nothing at all: the
-     * peer was told its registration succeeded and was never added to any
-     * dispatch list, so _notify_client_event() could not find it. The
-     * whole point of a default handler is to catch codes nobody asked for
-     * by name, so there is frequently no other registration to have
-     * created the entry first - and the very first client or tool to
-     * attach to a server necessarily finds the list empty. That client
-     * then silently received no default-routed event for its lifetime.
-     * The code != 0 path below has always created a missing entry; this
-     * one simply has to do the same. */
-    if (0 == ncodes) {
-        reginfo = NULL;
-        PMIX_LIST_FOREACH (rptr, &pmix_server_globals.events, pmix_regevents_info_t) {
-            if (PMIX_MAX_ERR_CONSTANT == rptr->code) {
-                reginfo = rptr;
-                break;
-            }
-        }
-        if (NULL == reginfo) {
-            reginfo = PMIX_NEW(pmix_regevents_info_t);
-            if (NULL == reginfo) {
-                rc = PMIX_ERR_NOMEM;
-                goto cleanup;
-            }
-            reginfo->code = PMIX_MAX_ERR_CONSTANT;
-            pmix_list_append(&pmix_server_globals.events, &reginfo->super);
-        }
-        prev = PMIX_NEW(pmix_peer_events_info_t);
-        if (NULL == prev) {
-            rc = PMIX_ERR_NOMEM;
-            goto cleanup;
-        }
-        PMIX_RETAIN(peer);
-        prev->peer = peer;
-        if (NULL != affected) {
-            PMIX_PROC_CREATE(prev->affected, naffected);
-            prev->naffected = naffected;
-            memcpy(prev->affected, affected, naffected * sizeof(pmix_proc_t));
-        }
-        pmix_list_append(&reginfo->peers, &prev->super);
-        /* fall through rather than returning here: a default handler must
-         * still be given any matching notification already sitting in the
-         * cache, and _check_cached_events has a dedicated arm for exactly
-         * this case (scd->codes == NULL matches every non-nondefault event).
-         * Returning early meant a default handler silently missed every
-         * event cached before it registered. The loops below are no-ops
-         * with ncodes == 0, and the trailing else branch schedules the
-         * cached-event check and returns PMIX_OPERATION_SUCCEEDED. */
-    }
-
-    /* store the event registration info so we can call the registered
-     * client when the server notifies the event */
-    for (n = 0; n < ncodes; n++) {
-        found = false;
-        PMIX_LIST_FOREACH (reginfo, &pmix_server_globals.events, pmix_regevents_info_t) {
-            if (PMIX_MAX_ERR_CONSTANT == reginfo->code) {
-                continue;
-            } else if (codes[n] == reginfo->code) {
-                found = true;
-                break;
-            }
-        }
-        if (found) {
-            /* found it - add this request */
-            prev = PMIX_NEW(pmix_peer_events_info_t);
-            if (NULL == prev) {
-                rc = PMIX_ERR_NOMEM;
-                goto cleanup;
-            }
-            PMIX_RETAIN(peer);
-            prev->peer = peer;
-            if (NULL != affected) {
-                PMIX_PROC_CREATE(prev->affected, naffected);
-                prev->naffected = naffected;
-                memcpy(prev->affected, affected, naffected * sizeof(pmix_proc_t));
-            }
-            prev->enviro_events = enviro_events;
-            pmix_list_append(&reginfo->peers, &prev->super);
-        } else {
-            /* if we get here, then we didn't find an existing registration for this code */
-            rptr = PMIX_NEW(pmix_regevents_info_t);
-            if (NULL == rptr) {
-                rc = PMIX_ERR_NOMEM;
-                goto cleanup;
-            }
-            rptr->code = codes[n];
-            pmix_list_append(&pmix_server_globals.events, &rptr->super);
-            prev = PMIX_NEW(pmix_peer_events_info_t);
-            if (NULL == prev) {
-                rc = PMIX_ERR_NOMEM;
-                goto cleanup;
-            }
-            PMIX_RETAIN(peer);
-            prev->peer = peer;
-            if (NULL != affected) {
-                PMIX_PROC_CREATE(prev->affected, naffected);
-                prev->naffected = naffected;
-                memcpy(prev->affected, affected, naffected * sizeof(pmix_proc_t));
-            }
-            prev->enviro_events = enviro_events;
-            pmix_list_append(&rptr->peers, &prev->super);
-        }
-    }
-
-    /* find the system codes in this request that our host is not
-     * already forwarding on behalf of another local registrant (or of
-     * ourselves) - those are the only ones we need to ask it about */
-    if (enviro_events) {
-        nactive = mark_activations(codes, ncodes);
-    }
-
-    /* if they asked for enviro events our host isn't already sending
-     * us, then call the local server */
-    if (enviro_events && 0 < nactive) {
-        /* if they don't support this, then we cannot do it */
-        if (NULL == pmix_host_server.register_events) {
-            rc = PMIX_ERR_NOT_SUPPORTED;
-            goto cleanup;
-        }
-        /* need to ensure the arrays don't go away until after the
-         * host RM is done with them */
-        scd = PMIX_NEW(pmix_setup_caddy_t);
-        if (NULL == scd) {
-            undo_activations(codes, nactive);
-            rc = PMIX_ERR_NOMEM;
-            goto cleanup;
-        }
-        PMIX_RETAIN(peer);
-        scd->peer = peer;
-        scd->codes = codes;
-        scd->ncodes = ncodes;
-        scd->info = info;
-        scd->ninfo = ninfo;
-        scd->opcbfunc = cbfunc;
-        scd->cbdata = cbdata;
-        /* only the codes needing activation were sorted to the front of
-         * the array, so that is all we hand the host - the caddy still
-         * carries the full array as the cached-event check needs it */
-        if (PMIX_SUCCESS
-            == (rc = pmix_host_server.register_events(scd->codes, nactive, scd->info,
-                                                      scd->ninfo, regevopcbfunc, scd))) {
-            /* the host will call us back when completed */
-            pmix_output_verbose(
-                2, pmix_server_globals.event_output,
-                "server register events: host server processing event registration");
-            if (NULL != affected) {
-                PMIX_PROC_FREE(affected, naffected);
-            }
-            return rc;
-        } else if (PMIX_OPERATION_SUCCEEDED == rc) {
-            /* we need to check cached notifications, but we want to ensure
-             * that occurs _after_ the client returns from registering the
-             * event handler in case the event is flagged for do_not_cache.
-             * Setup an event to fire after we return as that means it will
-             * occur after we send the registration response back to the client,
-             * thus guaranteeing that the client will get their registration
-             * callback prior to delivery of an event notification */
-            if (pmix_atomic_check_bool(&pmix_globals.progress_thread_stopped)) {
-                /* the caddy carries the codes and info arrays, but its
-                 * destructor does not free them - so release them here,
-                 * along with the affected array that was never attached */
-                if (NULL != scd->codes) {
-                    free(scd->codes);
-                }
-                if (NULL != scd->info) {
-                    PMIX_INFO_FREE(scd->info, scd->ninfo);
-                }
-                PMIX_RELEASE(scd);
-                if (NULL != affected) {
-                    PMIX_PROC_FREE(affected, naffected);
-                }
-                return PMIX_ERR_NOT_AVAILABLE;
-            }
-
-            /* scd->peer was already set (with its retain) above - taking a
-             * second reference here leaked one on every registration the
-             * host completed atomically */
-            scd->procs = affected;
-            scd->nprocs = naffected;
-            scd->opcbfunc = NULL;
-            scd->cbdata = NULL;
-            PMIX_THREADSHIFT(scd, _check_cached_events);
-            return rc;
-        } else {
-            /* host returned a genuine error and won't be calling the callback function */
-            pmix_output_verbose(2, pmix_server_globals.event_output,
-                                "server register events: host server reg events returned rc =%d",
-                                rc);
-            undo_activations(codes, nactive);
-            PMIX_RELEASE(scd);
-            goto cleanup;
-        }
-    } else {
-        rc = PMIX_OPERATION_SUCCEEDED;
-        /* we need to check cached notifications, but we want to ensure
-         * that occurs _after_ the client returns from registering the
-         * event handler in case the event is flagged for do_not_cache.
-         * Setup an event to fire after we return as that means it will
-         * occur after we send the registration response back to the client,
-         * thus guaranteeing that the client will get their registration
-         * callback prior to delivery of an event notification */
-        if (pmix_atomic_check_bool(&pmix_globals.progress_thread_stopped)) {
-            rc = PMIX_ERR_NOT_AVAILABLE;
-            goto cleanup;
-        }
-
-        scd = PMIX_NEW(pmix_setup_caddy_t);
-        PMIX_RETAIN(peer);
-        scd->peer = peer;
-        scd->codes = codes;
-        scd->ncodes = ncodes;
-        scd->procs = affected;
-        scd->nprocs = naffected;
-        scd->opcbfunc = NULL;
-        scd->cbdata = NULL;
-        PMIX_THREADSHIFT(scd, _check_cached_events);
-        if (NULL != info) {
-            PMIX_INFO_FREE(info, ninfo);
-        }
-        return rc;
-    }
-
-cleanup:
-    pmix_output_verbose(2, pmix_server_globals.event_output,
-                        "server register events: ninfo =%lu rc =%d", ninfo, rc);
-    if (NULL != info) {
-        PMIX_INFO_FREE(info, ninfo);
-    }
-    if (NULL != codes) {
-        free(codes);
-    }
-    if (NULL != affected) {
-        PMIX_PROC_FREE(affected, naffected);
-    }
-    return rc;
-}
-
-void pmix_server_deregister_events(pmix_peer_t *peer, pmix_buffer_t *buf)
-{
-    int32_t cnt;
-    pmix_status_t rc, code;
-    pmix_regevents_info_t *reginfo = NULL;
-    pmix_regevents_info_t *reginfo_next;
-    pmix_peer_events_info_t *prev;
-
-    pmix_output_verbose(2, pmix_server_globals.event_output,
-                        "%s recvd deregister events from %s",
-                        PMIX_NAME_PRINT(&pmix_globals.myid),
-                        PMIX_PEER_PRINT(peer));
-
-    /* unpack codes and process until done */
-    cnt = 1;
-    PMIX_BFROPS_UNPACK(rc, peer, buf, &code, &cnt, PMIX_STATUS);
-    while (PMIX_SUCCESS == rc) {
-        PMIX_LIST_FOREACH_SAFE (reginfo, reginfo_next, &pmix_server_globals.events,
-                                pmix_regevents_info_t) {
-            if (code == reginfo->code) {
-                /* found it - remove this peer from the list */
-                PMIX_LIST_FOREACH (prev, &reginfo->peers, pmix_peer_events_info_t) {
-                    if (prev->peer == peer) {
-                        /* found it */
-                        pmix_list_remove_item(&reginfo->peers, &prev->super);
-                        PMIX_RELEASE(prev);
-                        break;
-                    }
-                }
-                /* if nobody is left registered for this code - not the
-                 * server itself, nor any of our local clients - then
-                 * remove it and tell our host to stop forwarding it */
-                pmix_server_prune_reginfo(reginfo);
-            }
-        }
-        cnt = 1;
-        PMIX_BFROPS_UNPACK(rc, peer, buf, &code, &cnt, PMIX_STATUS);
-    }
-    if (PMIX_ERR_UNPACK_READ_PAST_END_OF_BUFFER != rc) {
-        PMIX_ERROR_LOG(rc);
-    }
-}
-
-static void local_cbfunc(pmix_status_t status, void *cbdata)
-{
-    pmix_notify_caddy_t *cd = (pmix_notify_caddy_t *) cbdata;
-
-    if (NULL != cd->cbfunc) {
-        cd->cbfunc(status, cd->cbdata);
-    }
-    PMIX_RELEASE(cd);
-}
-
-static void intermed_step(pmix_status_t status, void *cbdata)
-{
-    pmix_notify_caddy_t *cd = (pmix_notify_caddy_t *) cbdata;
-    pmix_status_t rc;
-
-    if (PMIX_SUCCESS != status) {
-        rc = status;
-        goto complete;
-    }
-
-    /* check the range directive - if it is LOCAL, then we are
-     * done. Otherwise, it needs to go up to our
-     * host for dissemination */
-    if (PMIX_RANGE_LOCAL == cd->range) {
-        rc = PMIX_SUCCESS;
-        goto complete;
-    }
-
-    /* pass it to our host RM for distribution */
-    if (NULL != pmix_host_server.notify_event) {
-        rc = pmix_host_server.notify_event(cd->status, &cd->source, cd->range,
-                                           cd->info, cd->ninfo,
-                                           local_cbfunc, cd);
-        if (PMIX_SUCCESS == rc) {
-            /* let the callback function respond for us */
-            return;
-        }
-        if (PMIX_OPERATION_SUCCEEDED == rc ||
-            PMIX_ERR_NOT_SUPPORTED == rc) {
-            rc = PMIX_SUCCESS; // local_cbfunc will not be called
-        }
-    } else {
-        rc = PMIX_SUCCESS; // local_cbfunc will not be called
-    }
-
-complete:
-    if (NULL != cd->cbfunc) {
-        cd->cbfunc(rc, cd->cbdata);
-    }
-    PMIX_RELEASE(cd);
-}
-
-/* Receive an event sent by the client library. Since it was sent
- * to us by one client, we have to both process it locally to ensure
- * we notify all relevant local clients AND (assuming a range other
- * than LOCAL) deliver to our host, requesting that they send it
- * to all peer servers in the current session */
-pmix_status_t pmix_server_event_recvd_from_client(pmix_peer_t *peer, pmix_buffer_t *buf,
-                                                  pmix_op_cbfunc_t cbfunc, void *cbdata)
-{
-    int32_t cnt;
-    pmix_status_t rc;
-    pmix_notify_caddy_t *cd;
-    size_t ninfo, n;
-
-    pmix_output_verbose(2, pmix_server_globals.event_output,
-                        "%s:%d recvd event notification from client %s:%d",
-                        pmix_globals.myid.nspace, pmix_globals.myid.rank, peer->info->pname.nspace,
-                        peer->info->pname.rank);
-
-    cd = PMIX_NEW(pmix_notify_caddy_t);
-    if (NULL == cd) {
-        return PMIX_ERR_NOMEM;
-    }
-    cd->cbfunc = cbfunc;
-    cd->cbdata = cbdata;
-    /* set the source */
-    PMIX_LOAD_PROCID(&cd->source, peer->info->pname.nspace, peer->info->pname.rank);
-
-    /* unpack status */
-    cnt = 1;
-    PMIX_BFROPS_UNPACK(rc, peer, buf, &cd->status, &cnt, PMIX_STATUS);
-    if (PMIX_SUCCESS != rc) {
-        PMIX_ERROR_LOG(rc);
-        goto exit;
-    }
-
-    /* unpack the range */
-    cnt = 1;
-    PMIX_BFROPS_UNPACK(rc, peer, buf, &cd->range, &cnt, PMIX_DATA_RANGE);
-    if (PMIX_SUCCESS != rc) {
-        PMIX_ERROR_LOG(rc);
-        goto exit;
-    }
-
-    /* unpack the info keys */
-    cnt = 1;
-    PMIX_BFROPS_UNPACK(rc, peer, buf, &ninfo, &cnt, PMIX_SIZE);
-    if (PMIX_SUCCESS != rc) {
-        PMIX_ERROR_LOG(rc);
-        goto exit;
-    }
-    cd->ninfo = ninfo + 1;
-    PMIX_INFO_CREATE(cd->info, cd->ninfo);
-    if (NULL == cd->info) {
-        rc = PMIX_ERR_NOMEM;
-        goto exit;
-    }
-    if (0 < ninfo) {
-        cnt = ninfo;
-        PMIX_BFROPS_UNPACK(rc, peer, buf, cd->info, &cnt, PMIX_INFO);
-        if (PMIX_SUCCESS != rc) {
-            PMIX_ERROR_LOG(rc);
-            goto exit;
-        }
-    }
-
-    /* if a local member is voluntarily leaving a group, account for its
-     * departure in any in-flight group construct/destruct collective for
-     * that group so the collective completes on the survivors rather than
-     * hanging - the deliberate cousin of a lost participant. We do this
-     * before the loop-detection check below because only the originating
-     * server (the one that received the leave from its client) sees this
-     * command; peer servers receive the relayed event with the
-     * PMIX_SERVER_INTERNAL_NOTIFY marker and bail out. */
-    if (PMIX_GROUP_LEFT == cd->status) {
-        char *grpid = NULL;
-        pmix_proc_t *affected = NULL;
-        for (n = 0; n < ninfo; n++) {
-            if (PMIX_CHECK_KEY(&cd->info[n], PMIX_GROUP_ID)) {
-                grpid = cd->info[n].value.data.string;
-            } else if (PMIX_CHECK_KEY(&cd->info[n], PMIX_EVENT_AFFECTED_PROC)) {
-                affected = cd->info[n].value.data.proc;
-            }
-        }
-        if (NULL != grpid && NULL != affected) {
-            pmix_server_grp_member_left(grpid, affected);
-        }
-    }
-
-    /* check to see if we already processed this event - it is possible
-     * that a local client "echoed" it back to us and we want to avoid
-     * a potential infinite loop */
-    for (n = 0; n < ninfo; n++) {
-        if (PMIX_CHECK_KEY(&cd->info[n], PMIX_SERVER_INTERNAL_NOTIFY)) {
-            /* yep, we did - so don't do it again! */
-            rc = PMIX_OPERATION_SUCCEEDED;
-            goto exit;
-        }
-    }
-
-    /* add an info object to mark that we recvd this internally */
-    PMIX_INFO_LOAD(&cd->info[cd->ninfo - 1], PMIX_SERVER_INTERNAL_NOTIFY, NULL, PMIX_BOOL);
-
-    /* process it */
-    rc = pmix_server_notify_client_of_event(cd->status, &cd->source, cd->range, cd->info,
-                                            cd->ninfo, intermed_step, cd);
-    if (PMIX_SUCCESS != rc) {
-        goto exit;
-    }
-    return rc;
-
-exit:
-    PMIX_RELEASE(cd);
-    return rc;
-}
-
-pmix_status_t pmix_server_query(pmix_peer_t *peer, pmix_buffer_t *buf,
-                                pmix_info_cbfunc_t cbfunc, void *cbdata)
-{
-    int32_t cnt;
-    pmix_status_t rc;
-    pmix_query_caddy_t *cd;
-
-     if (pmix_atomic_check_bool(&pmix_globals.progress_thread_stopped)) {
-        return PMIX_ERR_NOT_AVAILABLE;
-    }
-
-   pmix_output_verbose(2, pmix_server_globals.base_output,
-                        "recvd query from client");
-
-    cd = PMIX_NEW(pmix_query_caddy_t);
-    if (NULL == cd) {
-        return PMIX_ERR_NOMEM;
-    }
-    cd->cbfunc = cbfunc;
-    cd->cbdata = cbdata;  // this is the pmix_server_caddy_t we were given
-    /* unpack the number of queries */
-    cnt = 1;
-    PMIX_BFROPS_UNPACK(rc, peer, buf, &cd->nqueries, &cnt, PMIX_SIZE);
-    if (PMIX_SUCCESS != rc) {
-        PMIX_ERROR_LOG(rc);
-        PMIX_RELEASE(cd);
-        return rc;
-    }
-    /* unpack the queries */
-    if (0 < cd->nqueries) {
-        PMIX_QUERY_CREATE(cd->queries, cd->nqueries);
-        if (NULL == cd->queries) {
-            rc = PMIX_ERR_NOMEM;
-            PMIX_RELEASE(cd);
-            return rc;
-        }
-        cnt = cd->nqueries;
-        PMIX_BFROPS_UNPACK(rc, peer, buf, cd->queries, &cnt, PMIX_QUERY);
-        if (PMIX_SUCCESS != rc) {
-            PMIX_ERROR_LOG(rc);
-            PMIX_RELEASE(cd);
-            return rc;
-        }
-    }
-    PMIX_THREADSHIFT(cd, pmix_parse_localquery);
-    return PMIX_SUCCESS;
-}
-
-static void localcbfn(pmix_status_t status, void *cbdata)
-{
-    pmix_shift_caddy_t *cb = (pmix_shift_caddy_t *) cbdata;
-
-    if (NULL != cb->cbfunc.opcbfn) {
-        cb->cbfunc.opcbfn(status, cb->cbdata);
-    }
-    PMIX_RELEASE(cb);
-}
-
-pmix_status_t pmix_server_log(pmix_peer_t *peer, pmix_buffer_t *buf,
-                              pmix_op_cbfunc_t cbfunc,
-                              void *cbdata)
-{
-    int32_t cnt;
-    pmix_status_t rc;
-    pmix_shift_caddy_t *cd;
-    pmix_proc_t proc;
-    time_t timestamp;
-    PMIX_HIDE_UNUSED_PARAMS(cbfunc, cbdata);
-
-    pmix_output_verbose(2, pmix_plog_base_framework.framework_output,
-                        "pmix:server recvd log from client");
-
-    /* setup the requesting peer name */
-    pmix_strncpy(proc.nspace, peer->info->pname.nspace, PMIX_MAX_NSLEN);
-    proc.rank = peer->info->pname.rank;
-
-    cd = PMIX_NEW(pmix_shift_caddy_t);
-    if (NULL == cd) {
-        return PMIX_ERR_NOMEM;
-    }
-    if (PMIX_PEER_IS_EARLIER(peer, 3, 0, 0)) {
-        timestamp = -1;
-    } else {
-        /* unpack the timestamp */
-        cnt = 1;
-        PMIX_BFROPS_UNPACK(rc, peer, buf, &timestamp, &cnt, PMIX_TIME);
-        if (PMIX_SUCCESS != rc) {
-            PMIX_ERROR_LOG(rc);
-            goto exit;
-        }
-    }
-
-    /* unpack the number of data */
-    cnt = 1;
-    PMIX_BFROPS_UNPACK(rc, peer, buf, &cd->ninfo, &cnt, PMIX_SIZE);
-    if (PMIX_SUCCESS != rc) {
-        PMIX_ERROR_LOG(rc);
-        goto exit;
-    }
-    cnt = cd->ninfo;
-    PMIX_INFO_CREATE(cd->info, cd->ninfo);
-    /* the caddy owns this array and must release it when done */
-    cd->infocopy = true;
-    /* unpack the data */
-    if (0 < cnt) {
-        PMIX_BFROPS_UNPACK(rc, peer, buf, cd->info, &cnt, PMIX_INFO);
-        if (PMIX_SUCCESS != rc) {
-            PMIX_ERROR_LOG(rc);
-            goto exit;
-        }
-    }
-    /* unpack the number of directives */
-    cnt = 1;
-    PMIX_BFROPS_UNPACK(rc, peer, buf, &cd->ndirs, &cnt, PMIX_SIZE);
-    if (PMIX_SUCCESS != rc) {
-        PMIX_ERROR_LOG(rc);
-        goto exit;
-    }
-    cnt = cd->ndirs;
-    /* always add the source to the directives so we
-     * can tell downstream if this gets "upcalled" to
-     * our host for relay */
-    cd->ndirs = cnt + 1;
-    /* if a timestamp was sent, then we add it to the directives */
-    if (0 < timestamp) {
-        cd->ndirs++;
-        PMIX_INFO_CREATE(cd->directives, cd->ndirs);
-        PMIX_INFO_LOAD(&cd->directives[cnt], PMIX_LOG_SOURCE, &proc, PMIX_PROC);
-        PMIX_INFO_LOAD(&cd->directives[cnt + 1], PMIX_LOG_TIMESTAMP, &timestamp, PMIX_TIME);
-    } else {
-        PMIX_INFO_CREATE(cd->directives, cd->ndirs);
-        PMIX_INFO_LOAD(&cd->directives[cnt], PMIX_LOG_SOURCE, &proc, PMIX_PROC);
-    }
-    /* the caddy owns this array and must release it when done */
-    cd->dircopy = true;
-
-    /* unpack the directives */
-    if (0 < cnt) {
-        PMIX_BFROPS_UNPACK(rc, peer, buf, cd->directives, &cnt, PMIX_INFO);
-        if (PMIX_SUCCESS != rc) {
-            PMIX_ERROR_LOG(rc);
-            goto exit;
-        }
-    }
-
-    /* if we are not the gateway, or the host-only log param is set,
-     * then pass this up to the host for transfer to the gateway, if
-     * available */
-    if (!PMIX_PEER_IS_GATEWAY(pmix_globals.mypeer) || pmix_log_host_only) {
-        pmix_output_verbose(2, pmix_plog_base_framework.framework_output,
-                            "pmix:server not gateway");
-        cd->cbfunc.opcbfn = cbfunc;
-        cd->cbdata = cbdata;
-        if (NULL != pmix_host_server.log2) {
-            pmix_output_verbose(2, pmix_plog_base_framework.framework_output,
-                                "pmix:server using log2 upcall");
-            rc = pmix_host_server.log2(&proc, cd->info, cd->ninfo,
-                                       cd->directives, cd->ndirs,
-                                       localcbfn, (void *) cd);
-            if (PMIX_SUCCESS != rc) {
-                PMIX_RELEASE(cd);
-                return rc;
-            }
-
-        } else if (NULL != pmix_host_server.log) {
-            pmix_output_verbose(2, pmix_plog_base_framework.framework_output,
-                                "pmix:server using log upcall");
-            pmix_host_server.log(&proc, cd->info, cd->ninfo,
-                                 cd->directives, cd->ndirs,
-                                 localcbfn, (void *) cd);
-        } else {
-            // tell the client that we cannot do this
-            PMIX_RELEASE(cd);
-            return PMIX_ERR_NOT_AVAILABLE;
-        }
-        return PMIX_SUCCESS;
-    }
-
-    /* pass it down */
-    pmix_output_verbose(2, pmix_plog_base_framework.framework_output,
-                        "pmix:server processing locally");
-    rc = pmix_plog.log(&proc, cd->info, cd->ninfo, cd->directives, cd->ndirs);
-    if (PMIX_SUCCESS == rc) {
-        // mark that it was done atomically
-        rc = PMIX_OPERATION_SUCCEEDED;
-    }
-
-exit:
-    PMIX_RELEASE(cd);
-    return rc;
-}
-
-pmix_status_t pmix_server_alloc(pmix_peer_t *peer, pmix_buffer_t *buf,
-                                pmix_info_cbfunc_t cbfunc,
-                                void *cbdata)
-{
-    int32_t cnt;
-    pmix_status_t rc;
-    pmix_query_caddy_t *cd;
-    pmix_proc_t proc;
-    pmix_alloc_directive_t directive;
-
-    pmix_output_verbose(2, pmix_server_globals.base_output,
-                        "%s recvd allocate request from client %s",
-                        PMIX_NAME_PRINT(&pmix_globals.myid),
-                        PMIX_PEER_PRINT(peer));
-
-    if (NULL == pmix_host_server.allocate) {
-        return PMIX_ERR_NOT_SUPPORTED;
-    }
-
-    cd = PMIX_NEW(pmix_query_caddy_t);
-    if (NULL == cd) {
-        return PMIX_ERR_NOMEM;
-    }
-    cd->cbdata = cbdata;
-
-    /* unpack the directive */
-    cnt = 1;
-    PMIX_BFROPS_UNPACK(rc, peer, buf, &directive, &cnt, PMIX_ALLOC_DIRECTIVE);
-    if (PMIX_SUCCESS != rc) {
-        PMIX_ERROR_LOG(rc);
-        goto exit;
-    }
-
-    /* unpack the number of info objects */
-    cnt = 1;
-    PMIX_BFROPS_UNPACK(rc, peer, buf, &cd->ninfo, &cnt, PMIX_SIZE);
-    if (PMIX_SUCCESS != rc) {
-        PMIX_ERROR_LOG(rc);
-        goto exit;
-    }
-    /* unpack the info */
-    if (0 < cd->ninfo) {
-        PMIX_INFO_CREATE(cd->info, cd->ninfo);
-        cnt = cd->ninfo;
-        PMIX_BFROPS_UNPACK(rc, peer, buf, cd->info, &cnt, PMIX_INFO);
-        if (PMIX_SUCCESS != rc) {
-            PMIX_ERROR_LOG(rc);
-            goto exit;
-        }
-    }
-
-    /* setup the requesting peer name */
-    pmix_strncpy(proc.nspace, peer->info->pname.nspace, PMIX_MAX_NSLEN);
-    proc.rank = peer->info->pname.rank;
-
-    /* ask the host to execute the request */
-    rc = pmix_host_server.allocate(&proc, directive, cd->info, cd->ninfo, cbfunc, cd);
-    if (PMIX_SUCCESS != rc) {
-        goto exit;
-    }
-    return PMIX_SUCCESS;
-
-exit:
-    PMIX_RELEASE(cd);
-    return rc;
-}
-
-typedef struct {
-    pmix_list_item_t super;
-    pmix_epilog_t *epi;
-} pmix_srvr_epi_caddy_t;
-static PMIX_CLASS_INSTANCE(pmix_srvr_epi_caddy_t, pmix_list_item_t, NULL, NULL);
-
-pmix_status_t pmix_server_job_ctrl(pmix_peer_t *peer, pmix_buffer_t *buf,
-                                   pmix_info_cbfunc_t cbfunc,
-                                   void *cbdata)
-{
-    int32_t cnt, m;
-    pmix_status_t rc;
-    pmix_query_caddy_t *cd;
-    pmix_namespace_t *nptr, *tmp;
-    pmix_peer_t *pr;
-    pmix_proc_t proc;
-    size_t n;
-    bool recurse = false, leave_topdir = false, duplicate;
-    pmix_list_t cachedirs, cachefiles, ignorefiles, epicache;
-    pmix_srvr_epi_caddy_t *epicd = NULL;
-    pmix_cleanup_file_t *cf, *cf2, *cfptr;
-    pmix_cleanup_dir_t *cdir, *cdir2, *cdirptr;
-
-    pmix_output_verbose(2, pmix_server_globals.base_output,
-                        "recvd job control request from client");
-
-    if (NULL == pmix_host_server.job_control) {
-        return PMIX_ERR_NOT_SUPPORTED;
-    }
-
-    cd = PMIX_NEW(pmix_query_caddy_t);
-    if (NULL == cd) {
-        return PMIX_ERR_NOMEM;
-    }
-    cd->cbdata = cbdata;
-
-    /* construct every local list up front so that a single cleanup at the
-     * exit label covers all of them - the bad-param and no-memory paths in
-     * the directive scan below used to jump past their destructors and
-     * leak whatever cleanup entries had already been cached. Draining a
-     * list twice is harmless (PMIX_LIST_DESTRUCT leaves it re-initialized),
-     * so the mid-function destructs can stay where they are. */
-    PMIX_CONSTRUCT(&epicache, pmix_list_t);
-    PMIX_CONSTRUCT(&cachedirs, pmix_list_t);
-    PMIX_CONSTRUCT(&cachefiles, pmix_list_t);
-    PMIX_CONSTRUCT(&ignorefiles, pmix_list_t);
-
-    /* unpack the number of targets */
-    cnt = 1;
-    PMIX_BFROPS_UNPACK(rc, peer, buf, &cd->ntargets, &cnt, PMIX_SIZE);
-    if (PMIX_SUCCESS != rc) {
-        PMIX_ERROR_LOG(rc);
-        goto exit;
-    }
-    if (0 < cd->ntargets) {
-        PMIX_PROC_CREATE(cd->targets, cd->ntargets);
-        cnt = cd->ntargets;
-        PMIX_BFROPS_UNPACK(rc, peer, buf, cd->targets, &cnt, PMIX_PROC);
-        if (PMIX_SUCCESS != rc) {
-            PMIX_ERROR_LOG(rc);
-            goto exit;
-        }
-    }
-
-    /* check targets to find proper place to put any epilog requests */
-    if (NULL == cd->targets) {
-        epicd = PMIX_NEW(pmix_srvr_epi_caddy_t);
-        epicd->epi = &peer->nptr->epilog;
-        pmix_list_append(&epicache, &epicd->super);
-    } else {
-        for (n = 0; n < cd->ntargets; n++) {
-            /* find the nspace of this proc */
-            nptr = NULL;
-            PMIX_LIST_FOREACH (tmp, &pmix_globals.nspaces, pmix_namespace_t) {
-                if (0 == strcmp(tmp->nspace, cd->targets[n].nspace)) {
-                    nptr = tmp;
-                    break;
-                }
-            }
-            if (NULL == nptr) {
-                nptr = PMIX_NEW(pmix_namespace_t);
-                if (NULL == nptr) {
-                    rc = PMIX_ERR_NOMEM;
-                    goto exit;
-                }
-                nptr->nspace = strdup(cd->targets[n].nspace);
-                pmix_list_append(&pmix_globals.nspaces, &nptr->super);
-            }
-            /* if the rank is wildcard, then we use the epilog for the nspace */
-            if (PMIX_RANK_WILDCARD == cd->targets[n].rank) {
-                epicd = PMIX_NEW(pmix_srvr_epi_caddy_t);
-                epicd->epi = &nptr->epilog;
-                pmix_list_append(&epicache, &epicd->super);
-            } else {
-                /* we need to find the precise peer - we can only
-                 * do cleanup for a local client */
-                for (m = 0; m < pmix_server_globals.clients.size; m++) {
-                    pr = (pmix_peer_t *)pmix_pointer_array_get_item(&pmix_server_globals.clients, m);
-                    if (NULL == pr) {
-                        continue;
-                    }
-                    if (!PMIX_CHECK_NSPACE(pr->info->pname.nspace, cd->targets[n].nspace)) {
-                        continue;
-                    }
-                    if (pr->info->pname.rank == cd->targets[n].rank) {
-                        epicd = PMIX_NEW(pmix_srvr_epi_caddy_t);
-                        epicd->epi = &pr->epilog;
-                        pmix_list_append(&epicache, &epicd->super);
-                        break;
-                    }
-                }
-            }
-        }
-    }
-
-    /* unpack the number of info objects */
-    cnt = 1;
-    PMIX_BFROPS_UNPACK(rc, peer, buf, &cd->ninfo, &cnt, PMIX_SIZE);
-    if (PMIX_SUCCESS != rc) {
-        PMIX_ERROR_LOG(rc);
-        goto exit;
-    }
-    /* unpack the info */
-    if (0 < cd->ninfo) {
-        PMIX_INFO_CREATE(cd->info, cd->ninfo);
-        cnt = cd->ninfo;
-        PMIX_BFROPS_UNPACK(rc, peer, buf, cd->info, &cnt, PMIX_INFO);
-        if (PMIX_SUCCESS != rc) {
-            PMIX_ERROR_LOG(rc);
-            goto exit;
-        }
-    }
-
-    /* if this includes a request for post-termination cleanup, we handle
-     * that request ourselves */
-    cnt = 0; // track how many infos are cleanup related
-    for (n = 0; n < cd->ninfo; n++) {
-        if (PMIX_CHECK_KEY(&cd->info[n], PMIX_REGISTER_CLEANUP)) {
-            ++cnt;
-            if (PMIX_STRING != cd->info[n].value.type || NULL == cd->info[n].value.data.string) {
-                /* return an error */
-                rc = PMIX_ERR_BAD_PARAM;
-                goto exit;
-            }
-            cf = PMIX_NEW(pmix_cleanup_file_t);
-            if (NULL == cf) {
-                /* return an error */
-                rc = PMIX_ERR_NOMEM;
-                goto exit;
-            }
-            cf->path = strdup(cd->info[n].value.data.string);
-            pmix_list_append(&cachefiles, &cf->super);
-        } else if (PMIX_CHECK_KEY(&cd->info[n], PMIX_REGISTER_CLEANUP_DIR)) {
-            ++cnt;
-            if (PMIX_STRING != cd->info[n].value.type || NULL == cd->info[n].value.data.string) {
-                /* return an error */
-                rc = PMIX_ERR_BAD_PARAM;
-                goto exit;
-            }
-            cdir = PMIX_NEW(pmix_cleanup_dir_t);
-            if (NULL == cdir) {
-                /* return an error */
-                rc = PMIX_ERR_NOMEM;
-                goto exit;
-            }
-            cdir->path = strdup(cd->info[n].value.data.string);
-            pmix_list_append(&cachedirs, &cdir->super);
-        } else if (PMIX_CHECK_KEY(&cd->info[n], PMIX_CLEANUP_RECURSIVE)) {
-            recurse = PMIX_INFO_TRUE(&cd->info[n]);
-            ++cnt;
-        } else if (PMIX_CHECK_KEY(&cd->info[n], PMIX_CLEANUP_IGNORE)) {
-            if (PMIX_STRING != cd->info[n].value.type || NULL == cd->info[n].value.data.string) {
-                /* return an error */
-                rc = PMIX_ERR_BAD_PARAM;
-                goto exit;
-            }
-            cf = PMIX_NEW(pmix_cleanup_file_t);
-            if (NULL == cf) {
-                /* return an error */
-                rc = PMIX_ERR_NOMEM;
-                goto exit;
-            }
-            cf->path = strdup(cd->info[n].value.data.string);
-            pmix_list_append(&ignorefiles, &cf->super);
-            ++cnt;
-        } else if (PMIX_CHECK_KEY(&cd->info[n], PMIX_CLEANUP_LEAVE_TOPDIR)) {
-            leave_topdir = PMIX_INFO_TRUE(&cd->info[n]);
-            ++cnt;
-        }
-    }
-    if (0 < cnt) {
-        /* handle any ignore directives first */
-        PMIX_LIST_FOREACH (cf, &ignorefiles, pmix_cleanup_file_t) {
-            PMIX_LIST_FOREACH (epicd, &epicache, pmix_srvr_epi_caddy_t) {
-                /* scan the existing list of files for any duplicate */
-                duplicate = false;
-                PMIX_LIST_FOREACH (cf2, &epicd->epi->cleanup_files, pmix_cleanup_file_t) {
-                    if (0 == strcmp(cf2->path, cf->path)) {
-                        duplicate = true;
-                        break;
-                    }
-                }
-                if (!duplicate) {
-                    /* append a copy to the end of the list - cf itself is a
-                     * member of the local ignorefiles list (destructed below),
-                     * so appending &cf->super would put one item on two lists
-                     * and leave a dangling entry when ignorefiles is freed */
-                    cfptr = PMIX_NEW(pmix_cleanup_file_t);
-                    cfptr->path = strdup(cf->path);
-                    pmix_list_append(&epicd->epi->ignores, &cfptr->super);
-                }
-            }
-        }
-        PMIX_LIST_DESTRUCT(&ignorefiles);
-        /* now look at the directories */
-        PMIX_LIST_FOREACH (cdir, &cachedirs, pmix_cleanup_dir_t) {
-            PMIX_LIST_FOREACH (epicd, &epicache, pmix_srvr_epi_caddy_t) {
-                /* scan the existing list of directories for any duplicate */
-                duplicate = false;
-                PMIX_LIST_FOREACH (cdir2, &epicd->epi->cleanup_dirs, pmix_cleanup_dir_t) {
-                    if (0 == strcmp(cdir2->path, cdir->path)) {
-                        /* duplicate - check for difference in flags per RFC
-                         * precedence rules */
-                        if (!cdir->recurse && recurse) {
-                            cdir->recurse = recurse;
-                        }
-                        if (!cdir->leave_topdir && leave_topdir) {
-                            cdir->leave_topdir = leave_topdir;
-                        }
-                        duplicate = true;
-                        break;
-                    }
-                }
-                if (!duplicate) {
-                    /* check for conflict with ignore */
-                    PMIX_LIST_FOREACH (cf, &epicd->epi->ignores, pmix_cleanup_file_t) {
-                        if (0 == strcmp(cf->path, cdir->path)) {
-                            /* return an error */
-                            rc = PMIX_ERR_CONFLICTING_CLEANUP_DIRECTIVES;
-                            PMIX_LIST_DESTRUCT(&cachedirs);
-                            PMIX_LIST_DESTRUCT(&cachefiles);
-                            goto exit;
-                        }
-                    }
-                    /* append it to the end of the list */
-                    cdirptr = PMIX_NEW(pmix_cleanup_dir_t);
-                    cdirptr->path = strdup(cdir->path);
-                    cdirptr->recurse = recurse;
-                    cdirptr->leave_topdir = leave_topdir;
-                    pmix_list_append(&epicd->epi->cleanup_dirs, &cdirptr->super);
-                }
-            }
-        }
-        PMIX_LIST_DESTRUCT(&cachedirs);
-        PMIX_LIST_FOREACH (cf, &cachefiles, pmix_cleanup_file_t) {
-            PMIX_LIST_FOREACH (epicd, &epicache, pmix_srvr_epi_caddy_t) {
-                /* scan the existing list of files for any duplicate */
-                duplicate = false;
-                PMIX_LIST_FOREACH (cf2, &epicd->epi->cleanup_files, pmix_cleanup_file_t) {
-                    if (0 == strcmp(cf2->path, cf->path)) {
-                        duplicate = true;
-                        break;
-                    }
-                }
-                if (!duplicate) {
-                    /* check for conflict with ignore */
-                    PMIX_LIST_FOREACH (cf2, &epicd->epi->ignores, pmix_cleanup_file_t) {
-                        if (0 == strcmp(cf->path, cf2->path)) {
-                            /* return an error - cachedirs was already
-                             * destructed above, so only cachefiles remains */
-                            rc = PMIX_ERR_CONFLICTING_CLEANUP_DIRECTIVES;
-                            PMIX_LIST_DESTRUCT(&cachefiles);
-                            goto exit;
-                        }
-                    }
-                    /* append it to the end of the list */
-                    cfptr = PMIX_NEW(pmix_cleanup_file_t);
-                    cfptr->path = strdup(cf->path);
-                    pmix_list_append(&epicd->epi->cleanup_files, &cfptr->super);
-                }
-            }
-        }
-        PMIX_LIST_DESTRUCT(&cachefiles);
-        if (cnt == (int) cd->ninfo) {
-            /* nothing more to do */
-            rc = PMIX_OPERATION_SUCCEEDED;
-            goto exit;
-        }
-    }
-
-    /* setup the requesting peer name */
-    pmix_strncpy(proc.nspace, peer->info->pname.nspace, PMIX_MAX_NSLEN);
-    proc.rank = peer->info->pname.rank;
-
-    /* ask the host to execute the request */
-    if (PMIX_SUCCESS
-        != (rc = pmix_host_server.job_control(&proc, cd->targets, cd->ntargets, cd->info, cd->ninfo,
-                                              cbfunc, cd))) {
-        goto exit;
-    }
-    PMIX_LIST_DESTRUCT(&epicache);
-    PMIX_LIST_DESTRUCT(&cachedirs);
-    PMIX_LIST_DESTRUCT(&cachefiles);
-    PMIX_LIST_DESTRUCT(&ignorefiles);
-    return PMIX_SUCCESS;
-
-exit:
-    PMIX_RELEASE(cd);
-    PMIX_LIST_DESTRUCT(&epicache);
-    PMIX_LIST_DESTRUCT(&cachedirs);
-    PMIX_LIST_DESTRUCT(&cachefiles);
-    PMIX_LIST_DESTRUCT(&ignorefiles);
-    return rc;
-}
-
-pmix_status_t pmix_server_monitor(pmix_peer_t *peer, pmix_buffer_t *buf,
-                                  pmix_info_cbfunc_t cbfunc,
-                                  void *cbdata)
-{
-    int32_t cnt;
-    pmix_status_t rc;
-    pmix_cb_t *cb;
-
-    if (pmix_atomic_check_bool(&pmix_globals.progress_thread_stopped)) {
-        return PMIX_ERR_NOT_AVAILABLE;
-    }
-
-   pmix_output_verbose(2, pmix_server_globals.base_output,
-                        "recvd monitor request from client %s",
-                        PMIX_PEER_PRINT(peer));
-
-    cb = PMIX_NEW(pmix_cb_t);
-    if (NULL == cb) {
-        return PMIX_ERR_NOMEM;
-    }
-    PMIX_PROC_CREATE(cb->proc, 1);
-    PMIX_LOAD_PROCID(&cb->proc[0], peer->info->pname.nspace, peer->info->pname.rank);
-    cb->nprocs = 1;
-    cb->cbdata = cbdata;
-    cb->cbfunc.infofn = cbfunc;
-
-    /* unpack what is to be monitored */
-    cb->info = PMIx_Info_create(1);
-    cb->infocopy = true;
-    cnt = 1;
-    PMIX_BFROPS_UNPACK(rc, peer, buf, cb->info, &cnt, PMIX_INFO);
-    if (PMIX_SUCCESS != rc) {
-        PMIX_ERROR_LOG(rc);
-        goto exit;
-    }
-
-    /* unpack the error code */
-    cnt = 1;
-    PMIX_BFROPS_UNPACK(rc, peer, buf, &cb->status, &cnt, PMIX_STATUS);
-    if (PMIX_SUCCESS != rc) {
-        PMIX_ERROR_LOG(rc);
-        goto exit;
-    }
-
-    /* unpack the number of directives */
-    cnt = 1;
-    PMIX_BFROPS_UNPACK(rc, peer, buf, &cb->ndirs, &cnt, PMIX_SIZE);
-    if (PMIX_SUCCESS != rc) {
-        PMIX_ERROR_LOG(rc);
-        goto exit;
-    }
-    /* unpack the directives */
-    if (0 < cb->ndirs) {
-        PMIX_INFO_CREATE(cb->directives, cb->ndirs);
-        cnt = cb->ndirs;
-        PMIX_BFROPS_UNPACK(rc, peer, buf, cb->directives, &cnt, PMIX_INFO);
-        if (PMIX_SUCCESS != rc) {
-            PMIX_ERROR_LOG(rc);
-            goto exit;
-        }
-    }
-
-    // pass this over to be processed
-    PMIX_THREADSHIFT(cb, pmix_monitor_processing);
-    return PMIX_SUCCESS;
-
-exit:
-    PMIX_RELEASE(cb);
-    return rc;
-}
-
-pmix_status_t pmix_server_get_credential(pmix_peer_t *peer, pmix_buffer_t *buf,
-                                         pmix_credential_cbfunc_t cbfunc, void *cbdata)
-{
-    int32_t cnt;
-    pmix_status_t rc;
-    pmix_query_caddy_t *cd;
-    pmix_proc_t proc;
-
-    pmix_output_verbose(2, pmix_globals.debug_output, "recvd get credential request from client");
-
-    if (NULL == pmix_host_server.get_credential) {
-        return PMIX_ERR_NOT_SUPPORTED;
-    }
-
-    cd = PMIX_NEW(pmix_query_caddy_t);
-    if (NULL == cd) {
-        return PMIX_ERR_NOMEM;
-    }
-    cd->cbdata = cbdata;
-
-    /* unpack the number of directives */
-    cnt = 1;
-    PMIX_BFROPS_UNPACK(rc, peer, buf, &cd->ninfo, &cnt, PMIX_SIZE);
-    if (PMIX_SUCCESS != rc) {
-        PMIX_ERROR_LOG(rc);
-        goto exit;
-    }
-    /* unpack the directives */
-    if (0 < cd->ninfo) {
-        PMIX_INFO_CREATE(cd->info, cd->ninfo);
-        cnt = cd->ninfo;
-        PMIX_BFROPS_UNPACK(rc, peer, buf, cd->info, &cnt, PMIX_INFO);
-        if (PMIX_SUCCESS != rc) {
-            PMIX_ERROR_LOG(rc);
-            goto exit;
-        }
-    }
-
-    /* setup the requesting peer name */
-    pmix_strncpy(proc.nspace, peer->info->pname.nspace, PMIX_MAX_NSLEN);
-    proc.rank = peer->info->pname.rank;
-
-    /* ask the host to execute the request */
-    if (PMIX_SUCCESS
-        != (rc = pmix_host_server.get_credential(&proc, cd->info, cd->ninfo, cbfunc, cd))) {
-        goto exit;
-    }
-    return PMIX_SUCCESS;
-
-exit:
-    PMIX_RELEASE(cd);
-    return rc;
-}
-
-pmix_status_t pmix_server_validate_credential(pmix_peer_t *peer, pmix_buffer_t *buf,
-                                              pmix_validation_cbfunc_t cbfunc, void *cbdata)
-{
-    int32_t cnt;
-    pmix_status_t rc;
-    pmix_query_caddy_t *cd;
-    pmix_proc_t proc;
-
-    pmix_output_verbose(2, pmix_globals.debug_output,
-                        "recvd validate credential request from client");
-
-    if (NULL == pmix_host_server.validate_credential) {
-        return PMIX_ERR_NOT_SUPPORTED;
-    }
-
-    cd = PMIX_NEW(pmix_query_caddy_t);
-    if (NULL == cd) {
-        return PMIX_ERR_NOMEM;
-    }
-    cd->cbdata = cbdata;
-
-    /* unpack the credential */
-    cnt = 1;
-    PMIX_BFROPS_UNPACK(rc, peer, buf, &cd->bo, &cnt, PMIX_BYTE_OBJECT);
-    if (PMIX_SUCCESS != rc) {
-        PMIX_ERROR_LOG(rc);
-        goto exit;
-    }
-
-    /* unpack the number of directives */
-    cnt = 1;
-    PMIX_BFROPS_UNPACK(rc, peer, buf, &cd->ninfo, &cnt, PMIX_SIZE);
-    if (PMIX_SUCCESS != rc) {
-        PMIX_ERROR_LOG(rc);
-        goto exit;
-    }
-    /* unpack the directives */
-    if (0 < cd->ninfo) {
-        PMIX_INFO_CREATE(cd->info, cd->ninfo);
-        cnt = cd->ninfo;
-        PMIX_BFROPS_UNPACK(rc, peer, buf, cd->info, &cnt, PMIX_INFO);
-        if (PMIX_SUCCESS != rc) {
-            PMIX_ERROR_LOG(rc);
-            goto exit;
-        }
-    }
-
-    /* setup the requesting peer name */
-    pmix_strncpy(proc.nspace, peer->info->pname.nspace, PMIX_MAX_NSLEN);
-    proc.rank = peer->info->pname.rank;
-
-    /* ask the host to execute the request */
-    if (PMIX_SUCCESS
-        != (rc = pmix_host_server.validate_credential(&proc, &cd->bo, cd->info, cd->ninfo, cbfunc,
-                                                      cd))) {
-        goto exit;
-    }
-    return PMIX_SUCCESS;
-
-exit:
-    PMIX_RELEASE(cd);
-    return rc;
-}
-
-pmix_status_t pmix_server_iofreg(pmix_peer_t *peer, pmix_buffer_t *buf,
-                                 pmix_op_cbfunc_t cbfunc,
-                                 void *cbdata)
-{
-    int32_t cnt;
-    pmix_status_t rc;
-    pmix_setup_caddy_t *cd;
-    pmix_iof_req_t *req;
-    size_t refid;
-
-    pmix_output_verbose(2, pmix_server_globals.iof_output, "recvd IOF PULL request from client");
-
-    if (NULL == pmix_host_server.iof_pull) {
-        return PMIX_ERR_NOT_SUPPORTED;
-    }
-
-    cd = PMIX_NEW(pmix_setup_caddy_t);
-    if (NULL == cd) {
-        return PMIX_ERR_NOMEM;
-    }
-    cd->cbdata = cbdata; // this is the pmix_server_caddy_t
-
-    /* unpack the number of procs */
-    cnt = 1;
-    PMIX_BFROPS_UNPACK(rc, peer, buf, &cd->nprocs, &cnt, PMIX_SIZE);
-    if (PMIX_SUCCESS != rc) {
-        PMIX_ERROR_LOG(rc);
-        goto exit;
-    }
-    /* unpack the procs */
-    if (0 < cd->nprocs) {
-        PMIX_PROC_CREATE(cd->procs, cd->nprocs);
-        cnt = cd->nprocs;
-        PMIX_BFROPS_UNPACK(rc, peer, buf, cd->procs, &cnt, PMIX_PROC);
-        if (PMIX_SUCCESS != rc) {
-            PMIX_ERROR_LOG(rc);
-            goto exit;
-        }
-    }
-
-    /* unpack the number of directives */
-    cnt = 1;
-    PMIX_BFROPS_UNPACK(rc, peer, buf, &cd->ninfo, &cnt, PMIX_SIZE);
-    if (PMIX_SUCCESS != rc) {
-        PMIX_ERROR_LOG(rc);
-        goto exit;
-    }
-    /* unpack the directives */
-    if (0 < cd->ninfo) {
-        PMIX_INFO_CREATE(cd->info, cd->ninfo);
-        cnt = cd->ninfo;
-        PMIX_BFROPS_UNPACK(rc, peer, buf, cd->info, &cnt, PMIX_INFO);
-        if (PMIX_SUCCESS != rc) {
-            PMIX_ERROR_LOG(rc);
-            goto exit;
-        }
-    }
-
-    /* unpack the channels */
-    cnt = 1;
-    PMIX_BFROPS_UNPACK(rc, peer, buf, &cd->channels, &cnt, PMIX_IOF_CHANNEL);
-    if (PMIX_SUCCESS != rc) {
-        PMIX_ERROR_LOG(rc);
-        goto exit;
-    }
-
-    /* unpack their local reference id */
-    cnt = 1;
-    PMIX_BFROPS_UNPACK(rc, peer, buf, &refid, &cnt, PMIX_SIZE);
-    if (PMIX_SUCCESS != rc) {
-        PMIX_ERROR_LOG(rc);
-        goto exit;
-    }
-
-    /* add this peer/source/channel combination */
-    req = PMIX_NEW(pmix_iof_req_t);
-    if (NULL == req) {
-        rc = PMIX_ERR_NOMEM;
-        goto exit;
-    }
-    PMIX_RETAIN(peer);
-    req->requestor = peer;
-    req->nprocs = cd->nprocs;
-    if (0 < req->nprocs) {
-        PMIX_PROC_CREATE(req->procs, cd->nprocs);
-        memcpy(req->procs, cd->procs, req->nprocs * sizeof(pmix_proc_t));
-    }
-    req->channels = cd->channels;
-    req->remote_id = refid;
-    req->local_id = pmix_pointer_array_add(&pmix_globals.iof_requests, req);
-    cd->ncodes = req->local_id;
-
-    /* ask the host to execute the request */
-    rc = pmix_host_server.iof_pull(cd->procs, cd->nprocs,
-                                   cd->info, cd->ninfo,
-                                   cd->channels, cbfunc, cd);
-    if (PMIX_OPERATION_SUCCEEDED == rc) {
-        /* the host did it atomically - send the response. In
-         * this particular case, we can just use the cbfunc
-         * ourselves as it will threadshift and guarantee
-         * proper handling (i.e. that the refid will be
-         * returned in the response to the client) */
-        cbfunc(PMIX_SUCCESS, cd);
-        /* returning other than SUCCESS will cause the
-         * switchyard to release the cd object */
-        return PMIX_SUCCESS;
-    }
-    if (PMIX_SUCCESS == rc) {
-        /* the host accepted the request and now owns cd - it will release
-         * it when it invokes the callback. Falling through to the exit
-         * label here would free the caddy out from under the host and
-         * leave the callback operating on freed memory. */
-        return PMIX_SUCCESS;
-    }
-
-exit:
-    PMIX_RELEASE(cd);
-    return rc;
-}
-
-pmix_status_t pmix_server_iofdereg(pmix_peer_t *peer, pmix_buffer_t *buf,
-                                   pmix_op_cbfunc_t cbfunc,
-                                   void *cbdata)
-{
-    int32_t cnt;
-    pmix_status_t rc;
-    pmix_setup_caddy_t *cd;
-    pmix_iof_req_t *req;
-    size_t ninfo, refid;
-
-    pmix_output_verbose(2, pmix_server_globals.iof_output,
-                        "recvd IOF DEREGISTER from client");
-
-    if (NULL == pmix_host_server.iof_pull) {
-        return PMIX_ERR_NOT_SUPPORTED;
-    }
-
-    cd = PMIX_NEW(pmix_setup_caddy_t);
-    if (NULL == cd) {
-        return PMIX_ERR_NOMEM;
-    }
-    cd->cbdata = cbdata; // this is the pmix_server_caddy_t
-
-    /* unpack the number of directives */
-    cnt = 1;
-    PMIX_BFROPS_UNPACK(rc, peer, buf, &ninfo, &cnt, PMIX_SIZE);
-    if (PMIX_SUCCESS != rc) {
-        PMIX_ERROR_LOG(rc);
-        goto exit;
-    }
-    /* unpack the directives - note that we have to add one
-     * to tell the server to stop forwarding to this channel */
-    cd->ninfo = ninfo + 1;
-    PMIX_INFO_CREATE(cd->info, cd->ninfo);
-    if (0 < ninfo) {
-        cnt = ninfo;
-        PMIX_BFROPS_UNPACK(rc, peer, buf, cd->info, &cnt, PMIX_INFO);
-        if (PMIX_SUCCESS != rc) {
-            PMIX_ERROR_LOG(rc);
-            goto exit;
-        }
-    }
-    /* add the directive to stop forwarding */
-    PMIX_INFO_LOAD(&cd->info[ninfo], PMIX_IOF_STOP, NULL, PMIX_BOOL);
-
-    /* unpack the handler ID */
-    cnt = 1;
-    PMIX_BFROPS_UNPACK(rc, peer, buf, &refid, &cnt, PMIX_SIZE);
-    if (PMIX_SUCCESS != rc) {
-        PMIX_ERROR_LOG(rc);
-        goto exit;
-    }
-
-    /* get the referenced handler */
-    req = (pmix_iof_req_t *) pmix_pointer_array_get_item(&pmix_globals.iof_requests, refid);
-    if (NULL == req) {
-        /* already gone? */
-        rc = PMIX_ERR_NOT_FOUND;
-        goto exit;
-    }
-    pmix_pointer_array_set_item(&pmix_globals.iof_requests, refid, NULL);
-    PMIX_RELEASE(req);
-
-    /* tell the server to stop */
-    rc = pmix_host_server.iof_pull(cd->procs, cd->nprocs,
-                                   cd->info, cd->ninfo,
-                                   cd->channels, cbfunc, cd);
-    if (PMIX_OPERATION_SUCCEEDED == rc) {
-        /* the host did it atomically - send the response. In
-         * this particular case, we can just use the cbfunc
-         * ourselves as it will threadshift and guarantee
-         * proper handling */
-        cbfunc(PMIX_SUCCESS, cd);
-        /* returning other than SUCCESS will cause the
-         * switchyard to release the cd object */
-        return PMIX_SUCCESS;
-    }
-    if (PMIX_SUCCESS == rc) {
-        /* the host accepted the request and now owns cd - see the
-         * matching note in pmix_server_iofreg */
-        return PMIX_SUCCESS;
-    }
-
-exit:
-    /* mirror pmix_server_iofreg: on an async SUCCESS the host owns cd and
-     * releases it via the callback; on any error path here it was never
-     * handed off, so release it (previously it leaked on every error). */
-    PMIX_RELEASE(cd);
-    return rc;
-}
-
-static void stdcbfunc(pmix_status_t status, void *cbdata)
-{
-    pmix_setup_caddy_t *cd = (pmix_setup_caddy_t *) cbdata;
-
-    if (NULL != cd->opcbfunc) {
-        cd->opcbfunc(status, cd->cbdata);
-    }
-    if (NULL != cd->procs) {
-        PMIX_PROC_FREE(cd->procs, cd->nprocs);
-    }
-    if (NULL != cd->info) {
-        PMIX_INFO_FREE(cd->info, cd->ninfo);
-    }
-    if (NULL != cd->bo) {
-        PMIX_BYTE_OBJECT_FREE(cd->bo, 1);
-    }
-    PMIX_RELEASE(cd);
-}
-
-pmix_status_t pmix_server_iofstdin(pmix_peer_t *peer,
-                                   pmix_buffer_t *buf,
-                                   pmix_op_cbfunc_t cbfunc,
-                                   void *cbdata)
-{
-    int32_t cnt;
-    pmix_status_t rc;
-    pmix_proc_t source;
-    pmix_setup_caddy_t *cd;
-
-    pmix_output_verbose(2, pmix_server_globals.iof_output,
-                        "recvd stdin IOF data from tool %s",
-                        PMIX_PEER_PRINT(peer));
-
-    if (NULL == pmix_host_server.push_stdin) {
-        return PMIX_ERR_NOT_SUPPORTED;
-    }
-
-    cd = PMIX_NEW(pmix_setup_caddy_t);
-    if (NULL == cd) {
-        return PMIX_ERR_NOMEM;
-    }
-    cd->opcbfunc = cbfunc;
-    cd->cbdata = cbdata;
-
-    /* remember that this peer feeds us stdin - it is who IOF flow
-     * control has to reach if our host tells us to stop taking it */
-    peer->stdin_producer = true;
-
-    /* unpack the number of targets */
-    cnt = 1;
-    PMIX_BFROPS_UNPACK(rc, peer, buf, &cd->nprocs, &cnt, PMIX_SIZE);
-    if (PMIX_SUCCESS != rc) {
-        PMIX_ERROR_LOG(rc);
-        goto error;
-    }
-    if (0 < cd->nprocs) {
-        PMIX_PROC_CREATE(cd->procs, cd->nprocs);
-        if (NULL == cd->procs) {
-            rc = PMIX_ERR_NOMEM;
-            goto error;
-        }
-        cnt = cd->nprocs;
-        PMIX_BFROPS_UNPACK(rc, peer, buf, cd->procs, &cnt, PMIX_PROC);
-        if (PMIX_SUCCESS != rc) {
-            PMIX_ERROR_LOG(rc);
-            goto error;
-        }
-    }
-
-    /* unpack the number of directives */
-    cnt = 1;
-    PMIX_BFROPS_UNPACK(rc, peer, buf, &cd->ninfo, &cnt, PMIX_SIZE);
-    if (PMIX_SUCCESS != rc) {
-        PMIX_ERROR_LOG(rc);
-        goto error;
-    }
-    if (0 < cd->ninfo) {
-        PMIX_INFO_CREATE(cd->info, cd->ninfo);
-        if (NULL == cd->info) {
-            rc = PMIX_ERR_NOMEM;
-            goto error;
-        }
-        cnt = cd->ninfo;
-        PMIX_BFROPS_UNPACK(rc, peer, buf, cd->info, &cnt, PMIX_INFO);
-        if (PMIX_SUCCESS != rc) {
-            PMIX_ERROR_LOG(rc);
-            goto error;
-        }
-    }
-
-    /* unpack the data */
-    PMIX_BYTE_OBJECT_CREATE(cd->bo, 1);
-    if (NULL == cd->bo) {
-        rc = PMIX_ERR_NOMEM;
-        goto error;
-    }
-
-    cnt = 1;
-    PMIX_BFROPS_UNPACK(rc, peer, buf, cd->bo, &cnt, PMIX_BYTE_OBJECT);
-    if (PMIX_ERR_UNPACK_READ_PAST_END_OF_BUFFER == rc) {
-        /* it is okay for them to not send data */
-        PMIX_BYTE_OBJECT_FREE(cd->bo, 1);
-    } else if (PMIX_SUCCESS != rc) {
-        PMIX_ERROR_LOG(rc);
-        goto error;
-    }
-
-    /* pass the data to the host */
-    pmix_strncpy(source.nspace, peer->nptr->nspace, PMIX_MAX_NSLEN);
-    source.rank = peer->info->pname.rank;
-    if (PMIX_SUCCESS
-        != (rc = pmix_host_server.push_stdin(&source, cd->procs, cd->nprocs, cd->info, cd->ninfo,
-                                             cd->bo, stdcbfunc, cd))) {
-        if (PMIX_OPERATION_SUCCEEDED == rc) {
-            /* the host completed atomically and will not call back - invoke
-             * the callback ourselves (it sends the reply and releases cd)
-             * and tell the switchyard we are done. Previously this path
-             * leaked cd (its procs, info, and byte object). */
-            stdcbfunc(PMIX_SUCCESS, cd);
-            return PMIX_SUCCESS;
-        }
-        goto error;
-    }
-    return rc;
-
-error:
-    PMIX_RELEASE(cd);
-    return rc;
-}
-
-static void _fabric_response(int sd, short args, void *cbdata)
-{
-    pmix_query_caddy_t *qcd = (pmix_query_caddy_t *) cbdata;
-    PMIX_HIDE_UNUSED_PARAMS(sd, args);
-
-    /* qcd->cbfunc is the switchyard's fabric_cbfunc, which expects the
-     * query caddy - not the server caddy hanging off qcd->cbdata. Handing
-     * it qcd->cbdata made it read a pmix_server_caddy_t as a
-     * pmix_query_caddy_t and dereference the resulting garbage pointer.
-     * The callback chain owns qcd (and its info array) from here on, so
-     * we must not release it a second time. */
-    qcd->cbfunc(PMIX_SUCCESS, qcd->info, qcd->ninfo, qcd, NULL, NULL);
-}
-
-static void frcbfunc(pmix_status_t status, void *cbdata)
-{
-    pmix_query_caddy_t *qcd = (pmix_query_caddy_t *) cbdata;
-
-    PMIX_ACQUIRE_OBJECT(qcd);
-    qcd->status = status;
-    PMIX_POST_OBJECT(qcd);
-
-    PMIX_WAKEUP_THREAD(&qcd->lock);
-}
-/* we are being called from the PMIx server's switchyard function,
- * which means we are in an event and can access global data */
-pmix_status_t pmix_server_fabric_register(pmix_server_caddy_t *cd, pmix_buffer_t *buf,
-                                          pmix_info_cbfunc_t cbfunc)
-{
-    int32_t cnt;
-    pmix_status_t rc;
-    pmix_query_caddy_t *qcd = NULL;
-    pmix_proc_t proc;
-    pmix_fabric_t fabric;
-
-    if (pmix_atomic_check_bool(&pmix_globals.progress_thread_stopped)) {
-        return PMIX_ERR_NOT_AVAILABLE;
-    }
-
-    pmix_output_verbose(2, pmix_server_globals.base_output,
-                        "recvd register_fabric request from client");
-
-    qcd = PMIX_NEW(pmix_query_caddy_t);
-    if (NULL == qcd) {
-        return PMIX_ERR_NOMEM;
-    }
-    PMIX_RETAIN(cd);
-    qcd->cbfunc = cbfunc;
-    qcd->cbdata = cd;
-
-    /* unpack the number of directives */
-    cnt = 1;
-    PMIX_BFROPS_UNPACK(rc, cd->peer, buf, &qcd->ninfo, &cnt, PMIX_SIZE);
-    if (PMIX_SUCCESS != rc) {
-        PMIX_ERROR_LOG(rc);
-        goto exit;
-    }
-    /* unpack the directives */
-    if (0 < qcd->ninfo) {
-        PMIX_INFO_CREATE(qcd->info, qcd->ninfo);
-        cnt = qcd->ninfo;
-        PMIX_BFROPS_UNPACK(rc, cd->peer, buf, qcd->info, &cnt, PMIX_INFO);
-        if (PMIX_SUCCESS != rc) {
-            PMIX_ERROR_LOG(rc);
-            goto exit;
-        }
-    }
-
-    /* see if we support this request ourselves */
-    PMIX_FABRIC_CONSTRUCT(&fabric);
-    rc = pmix_pnet.register_fabric(&fabric, qcd->info, qcd->ninfo, frcbfunc, qcd);
-    if (PMIX_OPERATION_SUCCEEDED == rc) {
-        /* we need to respond, but we want to ensure
-         * that occurs _after_ the client returns from its API */
-        if (NULL != qcd->info) {
-            PMIX_INFO_FREE(qcd->info, qcd->ninfo);
-        }
-        qcd->info = fabric.info;
-        qcd->ninfo = fabric.ninfo;
-        PMIX_THREADSHIFT(qcd, _fabric_response);
-        return PMIX_SUCCESS;
-    } else if (PMIX_SUCCESS == rc) {
-        PMIX_WAIT_THREAD(&qcd->lock);
-        /* we need to respond, but we want to ensure
-         * that occurs _after_ the client returns from its API */
-        if (NULL != qcd->info) {
-            PMIX_INFO_FREE(qcd->info, qcd->ninfo);
-        }
-        qcd->info = fabric.info;
-        qcd->ninfo = fabric.ninfo;
-        PMIX_THREADSHIFT(qcd, _fabric_response);
-        return PMIX_SUCCESS;
-    }
-
-    /* if we don't internally support it, see if
-     * our host does */
-    if (NULL == pmix_host_server.fabric) {
-        rc = PMIX_ERR_NOT_SUPPORTED;
-        goto exit;
-    }
-
-    /* setup the requesting peer name */
-    PMIX_LOAD_PROCID(&proc, cd->peer->info->pname.nspace, cd->peer->info->pname.rank);
-
-    /* ask the host to execute the request */
-    if (PMIX_SUCCESS
-        != (rc = pmix_host_server.fabric(&proc, PMIX_FABRIC_REQUEST_INFO, qcd->info, qcd->ninfo,
-                                         cbfunc, qcd))) {
-        goto exit;
-    }
-    return PMIX_SUCCESS;
-
-exit:
-    /* every path reaching here failed before handing qcd to an async
-     * callback, so release it - and the reference it took on the server
-     * caddy, which the query caddy's destructor knows nothing about. The
-     * switchyard releases the caddy's own reference on our error return. */
-    if (NULL != qcd) {
-        PMIX_RELEASE(qcd);
-    }
-    PMIX_RELEASE(cd);
-    return rc;
-}
-
-pmix_status_t pmix_server_fabric_update(pmix_server_caddy_t *cd, pmix_buffer_t *buf,
-                                        pmix_info_cbfunc_t cbfunc)
-{
-    int32_t cnt;
-    size_t index;
-    pmix_status_t rc;
-    pmix_query_caddy_t *qcd;
-    pmix_proc_t proc;
-    pmix_fabric_t fabric;
-
-    if (pmix_atomic_check_bool(&pmix_globals.progress_thread_stopped)) {
-        return PMIX_ERR_NOT_AVAILABLE;
-    }
-
-    pmix_output_verbose(2, pmix_server_globals.base_output,
-                        "recvd update_fabric request from client");
-
-    qcd = PMIX_NEW(pmix_query_caddy_t);
-    if (NULL == qcd) {
-        return PMIX_ERR_NOMEM;
-    }
-    PMIX_RETAIN(cd);
-    qcd->cbfunc = cbfunc;
-    qcd->cbdata = cd;
-
-    /* unpack the fabric index */
-    cnt = 1;
-    PMIX_BFROPS_UNPACK(rc, cd->peer, buf, &index, &cnt, PMIX_SIZE);
-    if (PMIX_SUCCESS != rc) {
-        PMIX_ERROR_LOG(rc);
-        goto exit;
-    }
-
-    /* see if we support this request ourselves */
-    PMIX_FABRIC_CONSTRUCT(&fabric);
-    fabric.index = index;
-    rc = pmix_pnet.update_fabric(&fabric);
-    if (PMIX_SUCCESS == rc) {
-        /* we need to respond, but we want to ensure
-         * that occurs _after_ the client returns from its API */
-        if (NULL != qcd->info) {
-            PMIX_INFO_FREE(qcd->info, qcd->ninfo);
-        }
-        qcd->info = fabric.info;
-        qcd->ninfo = fabric.ninfo;
-        PMIX_THREADSHIFT(qcd, _fabric_response);
-        return rc;
-    }
-
-    /* if we don't internally support it, see if
-     * our host does */
-    if (NULL == pmix_host_server.fabric) {
-        rc = PMIX_ERR_NOT_SUPPORTED;
-        goto exit;
-    }
-
-    /* setup the requesting peer name */
-    pmix_strncpy(proc.nspace, cd->peer->info->pname.nspace, PMIX_MAX_NSLEN);
-    proc.rank = cd->peer->info->pname.rank;
-    /* add the index */
-    qcd->ninfo = 1;
-    PMIX_INFO_CREATE(qcd->info, qcd->ninfo);
-    PMIX_INFO_LOAD(&qcd->info[0], PMIX_FABRIC_INDEX, &index, PMIX_SIZE);
-
-    /* ask the host to execute the request */
-    if (PMIX_SUCCESS
-        != (rc = pmix_host_server.fabric(&proc, PMIX_FABRIC_UPDATE_INFO, qcd->info, qcd->ninfo,
-                                         cbfunc, qcd))) {
-        goto exit;
-    }
-    return PMIX_SUCCESS;
-
-exit:
-    /* every path reaching here failed before handing qcd to an async
-     * callback, so release it - and the reference it took on the server
-     * caddy, which the query caddy's destructor knows nothing about.
-     * The sibling pmix_server_fabric_register does the same. */
-    PMIX_RELEASE(qcd);
-    PMIX_RELEASE(cd);
-    return rc;
-}
-
-pmix_status_t pmix_server_device_dists(pmix_server_caddy_t *cd,
-                                       pmix_buffer_t *buf,
-                                       pmix_device_dist_cbfunc_t cbfunc)
-{
-    pmix_topology_t topo = {NULL, NULL};
-    pmix_cpuset_t cpuset = {NULL, NULL};
-    pmix_status_t rc;
-    pmix_device_distance_t *distances;
-    size_t ndist;
-    int cnt;
-    pmix_cb_t cb;
-    pmix_kval_t *kv;
-    pmix_proc_t proc;
-
-    /* unpack the topology they want us to use */
-    cnt = 1;
-    PMIX_BFROPS_UNPACK(rc, cd->peer, buf, &topo, &cnt, PMIX_TOPO);
-    if (PMIX_SUCCESS != rc) {
-        PMIX_ERROR_LOG(rc);
-        return rc;
-    }
-
-    /* unpack the cpuset */
-    cnt = 1;
-    PMIX_BFROPS_UNPACK(rc, cd->peer, buf, &cpuset, &cnt, PMIX_PROC_CPUSET);
-    if (PMIX_SUCCESS != rc) {
-        PMIX_ERROR_LOG(rc);
-        goto cleanup;
-    }
-
-    /* unpack any directives */
-    cnt = 1;
-    PMIX_BFROPS_UNPACK(rc, cd->peer, buf, &cd->ninfo, &cnt, PMIX_SIZE);
-    if (PMIX_SUCCESS != rc) {
-        PMIX_ERROR_LOG(rc);
-        goto cleanup;
-    }
-    if (0 < cd->ninfo) {
-        PMIX_INFO_CREATE(cd->info, cd->ninfo);
-        cnt = cd->ninfo;
-        PMIX_BFROPS_UNPACK(rc, cd->peer, buf, cd->info, &cnt, PMIX_INFO);
-        if (PMIX_SUCCESS != rc) {
-            PMIX_ERROR_LOG(rc);
-            goto cleanup;
-        }
-    }
-
-    /* if the provided topo is NULL, use my own */
-    if (NULL == topo.topology) {
-        if (NULL == pmix_globals.topology.topology) {
-            /* try to get it */
-            rc = pmix_hwloc_load_topology(&pmix_globals.topology);
-            if (PMIX_SUCCESS != rc) {
-                /* nothing we can do */
-                goto cleanup;
-            }
-        }
-        topo.topology = pmix_globals.topology.topology;
-    }
-
-    /* if the cpuset is NULL, see if we know the binding of the requesting process */
-    if (NULL == cpuset.bitmap) {
-        PMIX_CONSTRUCT(&cb, pmix_cb_t);
-        /* the cb destructor does not own cb.key, so point it at the
-         * string constant rather than a strdup that nothing would free */
-        cb.key = PMIX_CPUSET;
-        PMIX_LOAD_PROCID(&proc, cd->peer->info->pname.nspace, cd->peer->info->pname.rank);
-        cb.proc = &proc;
-        cb.scope = PMIX_LOCAL;
-        cb.copy = true;
-        PMIX_GDS_FETCH_KV(rc, cd->peer, &cb);
-        if (PMIX_SUCCESS != rc) {
-            PMIX_DESTRUCT(&cb);
-            goto cleanup;
-        }
-        if (0 == pmix_list_get_size(&cb.kvs)) {
-            /* a "successful" fetch that returned nothing - pmix_list_get_first
-             * would hand back the list sentinel, not NULL, so guard on the
-             * size before dereferencing it */
-            rc = PMIX_ERR_NOT_FOUND;
-            PMIX_DESTRUCT(&cb);
-            goto cleanup;
-        }
-        kv = (pmix_kval_t*)pmix_list_get_first(&cb.kvs);
-        if (NULL == kv->value || PMIX_STRING != kv->value->type) {
-            rc = PMIX_ERR_INVALID_VAL;
-            PMIX_DESTRUCT(&cb);
-            goto cleanup;
-        }
-        rc = pmix_hwloc_parse_cpuset_string(kv->value->data.string, &cpuset);
-        if (PMIX_SUCCESS != rc) {
-            PMIX_DESTRUCT(&cb);
-            goto cleanup;
-        }
-        PMIX_DESTRUCT(&cb);
-    }
-    /* compute the distances */
-    rc = pmix_hwloc_compute_distances(&topo, &cpuset, cd->info, cd->ninfo, &distances, &ndist);
-    if (PMIX_SUCCESS == rc) {
-        /* send the reply */
-        cbfunc(rc, distances, ndist, cd, NULL, NULL);
-        PMIX_DEVICE_DIST_FREE(distances, ndist);
-    }
-
-cleanup:
-    if (NULL != topo.topology &&
-        topo.topology != pmix_globals.topology.topology) {
-        pmix_hwloc_destruct_topology(&topo);
-    }
-    if (NULL != cpuset.bitmap) {
-        pmix_hwloc_destruct_cpuset(&cpuset);
-    }
-    if (NULL != cpuset.source) {
-        free(cpuset.source);
-    }
-    return rc;
-}
-
-pmix_status_t pmix_server_refresh_cache(pmix_server_caddy_t *cd,
-                                        pmix_buffer_t *buf,
-                                        pmix_op_cbfunc_t cbfunc)
-{
-    pmix_proc_t p;
-    char *nspace;
-    int cnt;
-    pmix_status_t rc;
-    pmix_cb_t cb;
-    pmix_buffer_t *pbkt;
-    pmix_kval_t *kv;
-    PMIX_HIDE_UNUSED_PARAMS(cbfunc);
-
-    // unpack the ID of the proc being requested
-    cnt = 1;
-    PMIX_BFROPS_UNPACK(rc, cd->peer, buf, &nspace, &cnt, PMIX_STRING);
-    if (PMIX_SUCCESS != rc) {
-        PMIX_ERROR_LOG(rc);
-        return rc;
-    }
-    PMIX_LOAD_NSPACE(p.nspace, nspace);
-    free(nspace);
-
-    cnt = 1;
-    PMIX_BFROPS_UNPACK(rc, cd->peer, buf, &p.rank, &cnt, PMIX_PROC_RANK);
-    if (PMIX_SUCCESS != rc) {
-        PMIX_ERROR_LOG(rc);
-        return rc;
-    }
-
-    /* retrieve the data for the specific rank they are asking about */
-    PMIX_CONSTRUCT(&cb, pmix_cb_t);
-    cb.proc = &p;
-    cb.scope = PMIX_REMOTE;
-    cb.copy = false;
-    PMIX_GDS_FETCH_KV(rc, pmix_globals.mypeer, &cb);
-    /* the fetch reports its result through the macro's status argument -
-     * cb.status is left at its constructed default, so packing that told
-     * the client SUCCESS even when nothing was found */
-    cb.status = rc;
-
-    // pack it up
-    pbkt = PMIX_NEW(pmix_buffer_t);
-    if (NULL == pbkt) {
-        PMIX_DESTRUCT(&cb);
-        return PMIX_ERR_NOMEM;
-    }
-    // start with the status
-    PMIX_BFROPS_PACK(rc, cd->peer, pbkt, &cb.status, 1, PMIX_STATUS);
-    if (PMIX_SUCCESS != rc) {
-        PMIX_ERROR_LOG(rc);
-        PMIX_RELEASE(pbkt);
-        PMIX_DESTRUCT(&cb);
-        return rc;
-    }
-    PMIX_LIST_FOREACH(kv, &cb.kvs, pmix_kval_t) {
-        PMIX_BFROPS_PACK(rc, cd->peer, pbkt, kv, 1, PMIX_KVAL);
-        if (PMIX_SUCCESS != rc) {
-            PMIX_ERROR_LOG(rc);
-            PMIX_RELEASE(pbkt);
-            PMIX_DESTRUCT(&cb);
-            return rc;
-        }
-    }
-    PMIX_DESTRUCT(&cb);
-
-    // send it back to the requestor
-    PMIX_SERVER_QUEUE_REPLY(rc, cd->peer, cd->hdr.tag, pbkt);
-    if (PMIX_SUCCESS != rc) {
-        PMIX_RELEASE(pbkt);
-    }
-
-    /* we queued the reply ourselves and return SUCCESS, so the switchyard
-     * will not release the caddy - we must do so here */
-    PMIX_RELEASE(cd);
-    return PMIX_SUCCESS;
-}
-
-pmix_status_t pmix_server_resblk(pmix_server_caddy_t *cd,
-                                 pmix_buffer_t *buf,
-                                 pmix_op_cbfunc_t cbfunc)
-{
-    int32_t cnt;
-    pmix_status_t rc;
-    pmix_setup_caddy_t *scd;
-    pmix_proc_t proc;
-    pmix_resource_block_directive_t directive;
-
-    pmix_output_verbose(2, pmix_server_globals.base_output,
-                        "%s recvd resource block request from client %s",
-                        PMIX_NAME_PRINT(&pmix_globals.myid),
-                        PMIX_PEER_PRINT(cd->peer));
-
-    if (NULL == pmix_host_server.resource_block) {
-        return PMIX_ERR_NOT_SUPPORTED;
-    }
-
-    scd = PMIX_NEW(pmix_setup_caddy_t);
-    if (NULL == scd) {
-        return PMIX_ERR_NOMEM;
-    }
-    scd->cbdata = cd;
-
-    /* unpack the directive */
-    cnt = 1;
-    PMIX_BFROPS_UNPACK(rc, cd->peer, buf, &directive, &cnt, PMIX_RESBLOCK_DIRECTIVE);
-    if (PMIX_SUCCESS != rc) {
-        PMIX_ERROR_LOG(rc);
-        goto exit;
-    }
-
-    // unpack the block name
-    cnt = 1;
-    PMIX_BFROPS_UNPACK(rc, cd->peer, buf, &scd->nspace, &cnt, PMIX_STRING);
-    if (PMIX_SUCCESS != rc) {
-        PMIX_ERROR_LOG(rc);
-        goto exit;
-    }
-
-    /* unpack the number of units */
-    cnt = 1;
-    PMIX_BFROPS_UNPACK(rc, cd->peer, buf, &scd->nunits, &cnt, PMIX_SIZE);
-    if (PMIX_SUCCESS != rc) {
-        PMIX_ERROR_LOG(rc);
-        goto exit;
-    }
-    /* unpack the resource units */
-    if (0 < scd->nunits) {
-        PMIX_RESOURCE_UNIT_CREATE(scd->units, scd->nunits);
-        cnt = scd->nunits;
-        PMIX_BFROPS_UNPACK(rc, cd->peer, buf, scd->units, &cnt, PMIX_RESOURCE_UNIT);
-        if (PMIX_SUCCESS != rc) {
-            PMIX_ERROR_LOG(rc);
-            goto exit;
-        }
-    }
-
-    /* unpack the number of info */
-    cnt = 1;
-    PMIX_BFROPS_UNPACK(rc, cd->peer, buf, &scd->ninfo, &cnt, PMIX_SIZE);
-    if (PMIX_SUCCESS != rc) {
-        PMIX_ERROR_LOG(rc);
-        goto exit;
-    }
-    /* unpack the info */
-    if (0 < scd->ninfo) {
-        PMIX_INFO_CREATE(scd->info, scd->ninfo);
-        cnt = scd->ninfo;
-        PMIX_BFROPS_UNPACK(rc, cd->peer, buf, scd->info, &cnt, PMIX_INFO);
-        if (PMIX_SUCCESS != rc) {
-            PMIX_ERROR_LOG(rc);
-            goto exit;
-        }
-    }
-
-    /* setup the requesting peer name */
-    pmix_strncpy(proc.nspace, cd->peer->info->pname.nspace, PMIX_MAX_NSLEN);
-    proc.rank = cd->peer->info->pname.rank;
-
-    /* ask the host to execute the request */
-    rc = pmix_host_server.resource_block(&proc, directive, scd->nspace,
-                                         scd->units, scd->nunits,
-                                         scd->info, scd->ninfo,
-                                         cbfunc, scd);
-    if (PMIX_SUCCESS != rc) {
-        goto exit;
-    }
-    return PMIX_SUCCESS;
-
-exit:
-    PMIX_RELEASE(scd);
-    return rc;
-}
-
-
-pmix_status_t pmix_server_session_ctrl(pmix_server_caddy_t *cd,
-                                       pmix_buffer_t *buf,
-                                       pmix_info_cbfunc_t cbfunc)
-{
-    int32_t cnt;
-    pmix_status_t rc;
-    pmix_shift_caddy_t *scd;
-    pmix_proc_t proc;
-
-    pmix_output_verbose(2, pmix_server_globals.base_output,
-                        "%s recvd session ctrl request from client %s",
-                        PMIX_NAME_PRINT(&pmix_globals.myid),
-                        PMIX_PEER_PRINT(cd->peer));
-
-    if (NULL == pmix_host_server.session_control) {
-        return PMIX_ERR_NOT_SUPPORTED;
-    }
-
-    scd = PMIX_NEW(pmix_shift_caddy_t);
-    if (NULL == scd) {
-        return PMIX_ERR_NOMEM;
-    }
-    scd->cbdata = cd;
-    scd->cbfunc.infocbfunc = cbfunc;
-
-    /* unpack the sessionID */
-    cnt = 1;
-    PMIX_BFROPS_UNPACK(rc, cd->peer, buf, &scd->sessionid, &cnt, PMIX_UINT32);
-    if (PMIX_SUCCESS != rc) {
-        PMIX_ERROR_LOG(rc);
-        goto exit;
-    }
-
-    /* unpack the number of info objects */
-    cnt = 1;
-    PMIX_BFROPS_UNPACK(rc, cd->peer, buf, &scd->ninfo, &cnt, PMIX_SIZE);
-    if (PMIX_SUCCESS != rc) {
-        PMIX_ERROR_LOG(rc);
-        goto exit;
-    }
-    /* unpack the info */
-    if (0 < scd->ninfo) {
-        PMIX_INFO_CREATE(scd->info, scd->ninfo);
-        cnt = scd->ninfo;
-        PMIX_BFROPS_UNPACK(rc, cd->peer, buf, scd->info, &cnt, PMIX_INFO);
-        if (PMIX_SUCCESS != rc) {
-            PMIX_ERROR_LOG(rc);
-            goto exit;
-        }
-    }
-
-    /* setup the requesting peer name */
-    pmix_strncpy(proc.nspace, cd->peer->info->pname.nspace, PMIX_MAX_NSLEN);
-    proc.rank = cd->peer->info->pname.rank;
-
-    /* ask the host to execute the request */
-    rc = pmix_host_server.session_control(&proc, scd->sessionid,
-                                          scd->info, scd->ninfo, cbfunc, scd);
-    if (PMIX_SUCCESS != rc) {
-        goto exit;
-    }
-    return PMIX_SUCCESS;
-
-exit:
-    PMIX_RELEASE(scd);
-    return rc;
-}
-
-/*****    INSTANCE SERVER LIBRARY CLASSES    *****/
-static void gcon(pmix_grpinfo_t *p)
-{
-    PMIX_LOAD_PROCID(&p->proc, NULL, PMIX_RANK_UNDEF);
-    PMIX_BYTE_OBJECT_CONSTRUCT(&p->blob);
-}
-static void gdes(pmix_grpinfo_t *p)
-{
-    PMIX_BYTE_OBJECT_DESTRUCT(&p->blob);
-}
-PMIX_CLASS_INSTANCE(pmix_grpinfo_t,
-                    pmix_list_item_t,
-                    gcon, gdes);
-
-static void tcon(pmix_server_trkr_t *t)
-{
-    t->event_active = false;
-    t->host_called = false;
-    t->local = true;
-    t->id = NULL;
-    memset(t->pname.nspace, 0, PMIX_MAX_NSLEN + 1);
-    t->pname.rank = PMIX_RANK_UNDEF;
-    t->pcs = NULL;
-    t->npcs = 0;
-    PMIX_CONSTRUCT(&t->nslist, pmix_list_t);
-    PMIX_CONSTRUCT_LOCK(&t->lock);
-    t->def_complete = false;
-    PMIX_CONSTRUCT(&t->local_cbs, pmix_list_t);
-    PMIX_CONSTRUCT(&t->departed, pmix_list_t);
-    t->nlocal = 0;
-    t->info = NULL;
-    t->ninfo = 0;
-    PMIX_CONSTRUCT(&t->grpinfo, pmix_list_t);
-    t->grpop = PMIX_GROUP_NONE;
-    /* this needs to be set explicitly */
-    t->collect_type = PMIX_COLLECT_INVALID;
-    t->modexcbfunc = NULL;
-    t->op_cbfunc = NULL;
-    t->info_cbfunc = NULL;
-    t->hybrid = false;
-    t->cbdata = NULL;
-}
-static void tdes(pmix_server_trkr_t *t)
-{
-    if (NULL != t->id) {
-        free(t->id);
-    }
-    PMIX_DESTRUCT_LOCK(&t->lock);
-    if (NULL != t->pcs) {
-        free(t->pcs);
-    }
-    PMIX_LIST_DESTRUCT(&t->local_cbs);
-    PMIX_LIST_DESTRUCT(&t->departed);
-    if (NULL != t->info) {
-        PMIX_INFO_FREE(t->info, t->ninfo);
-    }
-    PMIX_LIST_DESTRUCT(&t->grpinfo);
-    PMIX_LIST_DESTRUCT(&t->nslist);
-}
-PMIX_CLASS_INSTANCE(pmix_server_trkr_t,
-                    pmix_list_item_t,
-                    tcon, tdes);
-
-/* The single predicate for deciding whether a collective's local phase
- * is complete. The tracker definition must be complete (all participating
- * nspaces registered, so nlocal is final) AND every expected local
- * participant must be accounted for - either by having contributed (an
- * entry on local_cbs) or by having departed before contributing (an entry
- * on departed). A live participant that has not yet been heard from is on
- * neither list, so the sum stays below nlocal until it either contributes
- * or departs. The '>=' comparison is deliberate: it tolerates any
- * over-count from fork/exec'd clones without falsely reporting the
- * collective as incomplete. See docs/how-things-work/collectives. */
-bool pmix_server_trk_complete(pmix_server_trkr_t *trk)
-{
-    if (!trk->def_complete) {
-        return false;
-    }
-    return (pmix_list_get_size(&trk->local_cbs) +
-            pmix_list_get_size(&trk->departed)) >= trk->nlocal;
-}
-
-/* Record a collective's completion status in the tracker's info array. The
- * participant handlers seed a PMIX_LOCAL_COLLECTIVE_STATUS slot; locate it by
- * key rather than by position. It is the last element for the fence and
- * disconnect families, but connect appends per-participant endpoint info and
- * (for cross-namespace connects) job-level info AFTER that slot (see
- * pmix_server_connect), so a positional write to info[ninfo-1] would clobber
- * the appended info and leave the real status slot stale. Locating by key is
- * correct for every family and cannot underflow when info is unset. If the
- * slot is absent this is a no-op. Shared with the collective-status unit test
- * (test/unit/collective_status.c). */
-void pmix_server_set_collective_status(pmix_info_t *info, size_t ninfo,
-                                       pmix_status_t status)
-{
-    size_t n;
-
-    if (NULL == info) {
-        return;
-    }
-    for (n = 0; n < ninfo; n++) {
-        if (PMIX_CHECK_KEY(&info[n], PMIX_LOCAL_COLLECTIVE_STATUS)) {
-            info[n].value.data.status = status;
-            return;
-        }
-    }
-}
-
-static void cdcon(pmix_server_caddy_t *cd)
-{
-    memset(&cd->ev, 0, sizeof(pmix_event_t));
-    cd->event_active = false;
-    cd->trk = NULL;
-    cd->peer = NULL;
-    cd->info = NULL;
-    cd->ninfo = 0;
-    cd->query = NULL;
-    cd->key = NULL;
-}
-static void cddes(pmix_server_caddy_t *cd)
-{
-    if (cd->event_active) {
-        pmix_event_del(&cd->ev);
-    }
-    if (NULL != cd->trk) {
-        PMIX_RELEASE(cd->trk);
-    }
-    if (NULL != cd->peer) {
-        PMIX_RELEASE(cd->peer);
-    }
-    if (NULL != cd->info) {
-        PMIX_INFO_FREE(cd->info, cd->ninfo);
-    }
-    if (NULL != cd->query) {
-        PMIX_QUERY_FREE(cd->query, 1);
-    }
-    if (NULL != cd->key) {
-        free(cd->key);
-    }
-}
-PMIX_CLASS_INSTANCE(pmix_server_caddy_t,
-                    pmix_list_item_t,
-                    cdcon, cddes);
-
-static void scadcon(pmix_setup_caddy_t *p)
-{
-    p->peer = NULL;
-    memset(&p->proc, 0, sizeof(pmix_proc_t));
-    PMIX_CONSTRUCT_LOCK(&p->lock);
-    p->nspace = NULL;
-    p->codes = NULL;
-    p->ncodes = 0;
-    p->procs = NULL;
-    p->nprocs = 0;
-    p->apps = NULL;
-    p->napps = 0;
-    p->server_object = NULL;
-    p->nlocalprocs = 0;
-    p->info = NULL;
-    p->ninfo = 0;
-    p->units = NULL;
-    p->nunits = 0;
-    p->copied = false;
-    p->keys = NULL;
-    p->channels = PMIX_FWD_NO_CHANNELS;
-    pmix_iof_init_flags(&p->flags);
-    p->xoff = false;
-    p->bo = NULL;
-    p->nbo = 0;
-    p->cbfunc = NULL;
-    p->opcbfunc = NULL;
-    p->setupcbfunc = NULL;
-    p->lkcbfunc = NULL;
-    p->spcbfunc = NULL;
-    p->cbdata = NULL;
-}
-static void scaddes(pmix_setup_caddy_t *p)
-{
-    if (NULL != p->peer) {
-        PMIX_RELEASE(p->peer);
-    }
-    PMIX_PROC_FREE(p->procs, p->nprocs);
-    if (p->copied) {
-        if (NULL != p->info) {
-            PMIX_INFO_FREE(p->info, p->ninfo);
-        }
-        if (NULL != p->apps) {
-            PMIX_APP_FREE(p->apps, p->napps);
-        }
-    }
-    PMIX_RESOURCE_UNIT_FREE(p->units, p->nunits);
-    if (NULL != p->bo) {
-        PMIX_BYTE_OBJECT_FREE(p->bo, p->nbo);
-    }
-    PMIX_DESTRUCT_LOCK(&p->lock);
-    if (NULL != p->flags.file) {
-        free(p->flags.file);
-    }
-    if (NULL != p->flags.directory) {
-        free(p->flags.directory);
-    }
-}
-PMIX_EXPORT PMIX_CLASS_INSTANCE(pmix_setup_caddy_t,
-                                pmix_object_t,
-                                scadcon, scaddes);
-
-PMIX_CLASS_INSTANCE(pmix_trkr_caddy_t, pmix_object_t, NULL, NULL);
-
-static void dmcon(pmix_dmdx_remote_t *p)
-{
-    p->cd = NULL;
-}
-static void dmdes(pmix_dmdx_remote_t *p)
-{
-    if (NULL != p->cd) {
-        PMIX_RELEASE(p->cd);
-    }
-}
-PMIX_CLASS_INSTANCE(pmix_dmdx_remote_t,
-                    pmix_list_item_t,
-                    dmcon, dmdes);
-
-static void dmrqcon(pmix_dmdx_request_t *p)
-{
-    memset(&p->ev, 0, sizeof(pmix_event_t));
-    p->event_active = false;
-    p->lcd = NULL;
-    p->key = NULL;
-}
-static void dmrqdes(pmix_dmdx_request_t *p)
-{
-    if (p->event_active) {
-        pmix_event_del(&p->ev);
-    }
-    if (NULL != p->lcd) {
-        PMIX_RELEASE(p->lcd);
-    }
-    if (NULL != p->key) {
-        free(p->key);
-    }
-}
-PMIX_CLASS_INSTANCE(pmix_dmdx_request_t,
-                    pmix_list_item_t,
-                    dmrqcon, dmrqdes);
-
-static void lmcon(pmix_dmdx_local_t *p)
-{
-    memset(&p->proc, 0, sizeof(pmix_proc_t));
-    PMIX_CONSTRUCT(&p->loc_reqs, pmix_list_t);
-    p->info = NULL;
-    p->ninfo = 0;
-}
-static void lmdes(pmix_dmdx_local_t *p)
-{
-    if (NULL != p->info) {
-        PMIX_INFO_FREE(p->info, p->ninfo);
-    }
-    PMIX_LIST_DESTRUCT(&p->loc_reqs);
-}
-PMIX_CLASS_INSTANCE(pmix_dmdx_local_t,
-                    pmix_list_item_t,
-                    lmcon, lmdes);
-
-static void prevcon(pmix_peer_events_info_t *p)
-{
-    p->peer = NULL;
-    /* PMIX_NEW does not zero the allocation, and the default-handler
-     * registration path never assigns this one */
-    p->enviro_events = false;
-    p->affected = NULL;
-    p->naffected = 0;
-}
-static void prevdes(pmix_peer_events_info_t *p)
-{
-    if (NULL != p->peer) {
-        PMIX_RELEASE(p->peer);
-    }
-    if (NULL != p->affected) {
-        PMIX_PROC_FREE(p->affected, p->naffected);
-    }
-}
-PMIX_CLASS_INSTANCE(pmix_peer_events_info_t,
-                    pmix_list_item_t,
-                    prevcon, prevdes);
-
-static void regcon(pmix_regevents_info_t *p)
-{
-    PMIX_CONSTRUCT(&p->peers, pmix_list_t);
-    p->code = PMIX_MAX_ERR_CONSTANT;
-    p->nmine = 0;
-    p->active = false;
-}
-static void regdes(pmix_regevents_info_t *p)
-{
-    PMIX_LIST_DESTRUCT(&p->peers);
-}
-PMIX_CLASS_INSTANCE(pmix_regevents_info_t,
-                    pmix_list_item_t,
-                    regcon, regdes);
-
-static void ilcon(pmix_inventory_rollup_t *p)
-{
-    PMIX_CONSTRUCT_LOCK(&p->lock);
-    p->status = PMIX_SUCCESS;
-    p->requests = 0;
-    p->replies = 0;
-    PMIX_CONSTRUCT(&p->payload, pmix_list_t);
-    p->info = NULL;
-    p->ninfo = 0;
-    p->cbfunc = NULL;
-    p->infocbfunc = NULL;
-    p->opcbfunc = NULL;
-    p->cbdata = NULL;
-}
-static void ildes(pmix_inventory_rollup_t *p)
-{
-    PMIX_DESTRUCT_LOCK(&p->lock);
-    PMIX_LIST_DESTRUCT(&p->payload);
-}
-PMIX_CLASS_INSTANCE(pmix_inventory_rollup_t,
-                    pmix_object_t,
-                    ilcon, ildes);
-
-PMIX_CLASS_INSTANCE(pmix_group_caddy_t,
-                    pmix_list_item_t,
-                    NULL, NULL);
-
-static void iocon(pmix_iof_cache_t *p)
-{
-    p->bo = NULL;
-    p->info = NULL;
-    p->ninfo = 0;
-}
-static void iodes(pmix_iof_cache_t *p)
-{
-    PMIX_BYTE_OBJECT_FREE(p->bo, 1); // macro protects against NULL
-    if (0 < p->ninfo) {
-        PMIX_INFO_FREE(p->info, p->ninfo);
-    }
-}
-PMIX_CLASS_INSTANCE(pmix_iof_cache_t,
-                    pmix_list_item_t,
-                    iocon, iodes);
-
-static void pscon(pmix_pset_t *p)
-{
-    p->name = NULL;
-    p->members = NULL;
-    p->nmembers = 0;
-}
-static void psdes(pmix_pset_t *p)
-{
-    if (NULL != p->name) {
-        free(p->name);
-    }
-    if (NULL != p->members) {
-        free(p->members);
-    }
-}
-PMIX_CLASS_INSTANCE(pmix_pset_t,
-                    pmix_list_item_t,
-                    pscon, psdes);
